Split api handlers
