+++ src/llm/BUILD
@@ -49,6 +49,25 @@ mediapipe_proto_library(
 )
 
 ovms_cc_library(
+    name = "openai_api_handler",
+    hdrs = ["apis/openai_api_handler.hpp"],
+    srcs = ["apis/openai_api_handler.cpp"],
+    deps = [
+        "@com_github_tencent_rapidjson//:rapidjson",
+        "@mediapipe//mediapipe/framework:calculator_framework", # required for absl status
+        "//src:libovmsprofiler",
+        "//third_party:curl",
+        "//src:image_conversion",
+        "//src:libovmsfilesystem",
+        "@stb//:image",
+        ":openai_request",
+        ":output_parsers",
+        "//third_party:genai",],
+    visibility = ["//visibility:public"],
+    additional_copts = COPTS_PYTHON
+)
+
+ovms_cc_library(
     name = "openai_completions_api_handler",
     hdrs = ["apis/openai_completions.hpp", "apis/openai_json_response.hpp"],
     srcs = ["apis/openai_completions.cpp"],
@@ -60,6 +79,27 @@ ovms_cc_library(
         "//src:image_conversion",
         "//src/filesystem:libovmsfilesystem",
         "@stb//:image",
+        ":openai_api_handler",
+        ":openai_request",
+        ":output_parsers",
+        "//third_party:genai",],
+    visibility = ["//visibility:public"],
+    additional_copts = COPTS_PYTHON
+)
+
+ovms_cc_library(
+    name = "openai_responses_handler",
+    hdrs = ["apis/openai_responses.hpp"],
+    srcs = ["apis/openai_responses.cpp"],
+    deps = [
+        "@com_github_tencent_rapidjson//:rapidjson",
+        "@mediapipe//mediapipe/framework:calculator_framework", # required for absl status
+        "//src:libovmsprofiler",
+        "//third_party:curl",
+        "//src:image_conversion",
+        "//src:libovmsfilesystem",
+        "@stb//:image",
+        ":openai_api_handler",
         ":openai_request",
         ":output_parsers",
         "//third_party:genai",],
@@ -237,6 +277,7 @@ ovms_cc_library(
         "//src/tokenize:tokenize_parser",
         "llmcalculator_cc_proto",
         ":openai_completions_api_handler",
+        ":openai_responses_handler",
         ":generation_config_builders",
         "//src:httppayload",
         "//src:libhttpclientconnection",
