+++ src/llm/apis/openai_completions.cpp
@@ -23,7 +23,12 @@
 #include "src/port/rapidjson_stringbuffer.hpp"
 #include "src/port/rapidjson_writer.hpp"
 #include <set>
+#include <string>
 #include <string.h>
+#include <vector>
+
+#include <openvino/genai/llm_pipeline.hpp>
+#include <openvino/genai/visual_language/pipeline.hpp>
 
 #include <fmt/ranges.h>
 
@@ -38,434 +43,34 @@
 #include "absl/strings/str_cat.h"
 #pragma warning(pop)
 
-#include <curl/curl.h>
 #include <regex>
-#include "../../image_conversion.hpp"  // TODO: Rename to stbi_conversions?
+#include "../../image_conversion.hpp"
 
 using namespace rapidjson;
 
 namespace ovms {
 
-constexpr size_t DEFAULT_MAX_STOP_WORDS = 16;  // same as deep-seek
-constexpr std::string_view BASE64_PREFIX = "base64,";
-constexpr int64_t MAX_IMAGE_SIZE_BYTES = 20000000;  // 20MB
-
-namespace {
-
-ov::genai::JsonContainer rapidJsonValueToJsonContainer(const rapidjson::Value& value) {
-    if (value.IsNull()) {
-        return ov::genai::JsonContainer(nullptr);
+static bool hasToolCallsInStreamingDelta(const rapidjson::Document& delta) {
+    if (!delta.HasMember("delta") || !delta["delta"].IsObject()) {
+        return false;
     }
-    if (value.IsBool()) {
-        return ov::genai::JsonContainer(value.GetBool());
-    }
-    if (value.IsInt()) {
-        return ov::genai::JsonContainer(value.GetInt());
-    }
-    if (value.IsUint()) {
-        return ov::genai::JsonContainer(static_cast<int64_t>(value.GetUint()));
-    }
-    if (value.IsInt64()) {
-        return ov::genai::JsonContainer(value.GetInt64());
-    }
-    if (value.IsUint64()) {
-        auto uintValue = value.GetUint64();
-        if (uintValue <= static_cast<uint64_t>(std::numeric_limits<int64_t>::max())) {
-            return ov::genai::JsonContainer(static_cast<int64_t>(uintValue));
-        }
-        return ov::genai::JsonContainer(static_cast<double>(uintValue));
-    }
-    if (value.IsDouble()) {
-        return ov::genai::JsonContainer(value.GetDouble());
-    }
-    if (value.IsString()) {
-        return ov::genai::JsonContainer(std::string(value.GetString(), value.GetStringLength()));
-    }
-    if (value.IsArray()) {
-        ov::genai::JsonContainer arrayContainer = ov::genai::JsonContainer::array();
-        for (const auto& item : value.GetArray()) {
-            arrayContainer.push_back(rapidJsonValueToJsonContainer(item));
-        }
-        return arrayContainer;
-    }
-    if (value.IsObject()) {
-        ov::genai::JsonContainer objectContainer = ov::genai::JsonContainer::object();
-        for (auto member = value.MemberBegin(); member != value.MemberEnd(); ++member) {
-            const std::string key(member->name.GetString(), member->name.GetStringLength());
-            objectContainer[key] = rapidJsonValueToJsonContainer(member->value);
-        }
-        return objectContainer;
-    }
-    throw std::invalid_argument("Unsupported JSON value type");
+    const auto& deltaObj = delta["delta"];
+    return deltaObj.HasMember("tool_calls") && deltaObj["tool_calls"].IsArray();
 }
 
-}  // namespace
+// --- Request parsing ---
 
-void OpenAIChatCompletionsHandler::serializeResponsesToolChoice(Writer<StringBuffer>& writer) const {
-    writer.String("tool_choice");
-    if (request.toolChoice.empty()) {
-        writer.String("auto");
-    } else if (request.toolChoice == "auto" || request.toolChoice == "none" || request.toolChoice == "required") {
-        writer.String(request.toolChoice.c_str());
-    } else {
-        writer.StartObject();
-        writer.String("type");
-        writer.String("function");
-        writer.String("name");
-        writer.String(request.toolChoice.c_str());
-        writer.EndObject();
-    }
-}
+absl::Status OpenAIChatCompletionsHandler::parseRequest(std::optional<uint32_t> maxTokensLimit, uint32_t bestOfLimit, std::optional<uint32_t> maxModelLength,
+    std::optional<std::string> allowedLocalMediaPath, std::optional<std::vector<std::string>> allowedMediaDomains) {
+    absl::Status status = parseCommonPart(maxTokensLimit, bestOfLimit, maxModelLength);
+    if (status != absl::OkStatus())
+        return status;
+    if (endpoint == Endpoint::COMPLETIONS)
+        status = parseCompletionsPart();
+    else
+        status = parseChatCompletionsPart(maxTokensLimit, allowedLocalMediaPath, allowedMediaDomains);
 
-void OpenAIChatCompletionsHandler::serializeResponsesTools(Writer<StringBuffer>& writer) const {
-    writer.String("tools");
-    writer.StartArray();
-    for (const auto& [toolName, toolSchemaWrapper] : request.toolNameSchemaMap) {
-        writer.StartObject();
-        writer.String("type");
-        writer.String("function");
-        writer.String("name");
-        writer.String(toolName.c_str());
-        writer.String("parameters");
-        writer.RawValue(toolSchemaWrapper.stringRepr.c_str(), toolSchemaWrapper.stringRepr.size(), rapidjson::kObjectType);
-        writer.EndObject();
-    }
-    writer.EndArray();
-}
-
-void OpenAIChatCompletionsHandler::serializeResponsesResponseObject(Writer<StringBuffer>& writer, const std::string& responseId, int64_t createdAt,
-    const std::string& status, const std::string& fullOutputText, bool includeUsage,
-    const std::optional<std::string>& incompleteReason, const std::optional<std::string>& errorMessage, ResponsesErrorCode errorCode) const {
-    writer.StartObject();
-    writer.String("id");
-    writer.String(responseId.c_str());
-    writer.String("object");
-    writer.String("response");
-    writer.String("created_at");
-    writer.Int64(createdAt);
-    if (status == "completed") {
-        const auto completedAt = std::chrono::duration_cast<std::chrono::seconds>(std::chrono::system_clock::now().time_since_epoch()).count();
-        writer.String("completed_at");
-        writer.Int64(completedAt);
-    }
-    if (incompleteReason.has_value()) {
-        writer.String("incomplete_details");
-        writer.StartObject();
-        writer.String("reason");
-        writer.String(incompleteReason.value().c_str());
-        writer.EndObject();
-    }
-    writer.String("error");
-    if (errorMessage.has_value()) {
-        writer.StartObject();
-        writer.String("code");
-        writer.String(responsesErrorCodeToString(errorCode));
-        writer.String("message");
-        writer.String(errorMessage.value().c_str());
-        writer.EndObject();
-    } else {
-        writer.Null();
-    }
-    writer.String("model");
-    writer.String(request.model.c_str());
-    writer.String("status");
-    writer.String(status.c_str());
-
-    writer.String("parallel_tool_calls");
-    writer.Bool(true);
-    // TODO: previous_response_id not supported
-    writer.String("store");
-    writer.Bool(true);
-    // TODO: temperature are only included when explicitly provided in the request
-    if (request.temperature.has_value()) {
-        writer.String("temperature");
-        writer.Double(static_cast<double>(request.temperature.value()));
-    }
-    writer.String("text");
-    writer.StartObject();
-    writer.String("format");
-    writer.StartObject();
-    writer.String("type");
-    writer.String("text");
-    writer.EndObject();
-    writer.EndObject();
-    serializeResponsesToolChoice(writer);
-    serializeResponsesTools(writer);
-    // TODO: top_p are only included when explicitly provided in the request
-    if (request.topP.has_value()) {
-        writer.String("top_p");
-        writer.Double(static_cast<double>(request.topP.value()));
-    }
-    writer.String("truncation");
-    writer.String("disabled");
-    // TODO: user not supported
-    writer.String("metadata");
-    writer.StartObject();
-    writer.EndObject();
-
-    if (request.maxTokens.has_value()) {
-        writer.String("max_output_tokens");
-        writer.Uint64(static_cast<uint64_t>(request.maxTokens.value()));
-    }
-
-    writer.String("output");
-    writer.StartArray();
-    // Include reasoning output item if reasoning was produced during streaming
-    if (!responsesState.reasoningText.empty()) {
-        writer.StartObject();
-        writer.String("id");
-        writer.String("rs-0");
-        writer.String("type");
-        writer.String("reasoning");
-        writer.String("summary");
-        writer.StartArray();
-        writer.StartObject();
-        writer.String("type");
-        writer.String("summary_text");
-        writer.String("text");
-        writer.String(responsesState.reasoningText.c_str());
-        writer.EndObject();
-        writer.EndArray();
-        writer.EndObject();
-    }
-    // Include function_call output items if tool calls were produced during streaming
-    for (const auto& toolCall : responsesState.toolCalls) {
-        writer.StartObject();
-        writer.String("id");
-        writer.String(toolCall.id.c_str());
-        writer.String("type");
-        writer.String("function_call");
-        writer.String("status");
-        writer.String(status.c_str());
-        writer.String("call_id");
-        writer.String(toolCall.id.c_str());
-        writer.String("name");
-        writer.String(toolCall.name.c_str());
-        writer.String("arguments");
-        writer.String(toolCall.arguments.c_str());
-        writer.EndObject();
-    }
-    {
-        writer.StartObject();
-        writer.String("id");
-        writer.String("msg-0");
-        writer.String("type");
-        writer.String("message");
-        writer.String("role");
-        writer.String("assistant");
-        writer.String("status");
-        writer.String(status.c_str());
-        writer.String("content");
-        writer.StartArray();
-        serializeOutputTextPart(writer, fullOutputText);
-        writer.EndArray();
-        writer.EndObject();
-    }
-    writer.EndArray();
-
-    if (includeUsage) {
-        writer.String("usage");
-        writer.StartObject();
-        writer.String("input_tokens");
-        writer.Uint64(static_cast<uint64_t>(usage.promptTokens));
-        // TODO: input_tokens_details.cached_tokens not supported
-        writer.String("output_tokens");
-        writer.Uint64(static_cast<uint64_t>(usage.completionTokens));
-        // TODO: output_tokens_details.reasoning_tokens not supported
-        writer.String("total_tokens");
-        writer.Uint64(static_cast<uint64_t>(usage.calculateTotalTokens()));
-        writer.EndObject();
-    }
-
-    writer.EndObject();
-}
-
-void OpenAIChatCompletionsHandler::serializeResponsesOutputItem(Writer<StringBuffer>& writer, const std::string& outputItemId,
-    const std::string& text, const std::string& status) {
-    writer.StartObject();
-    writer.String("id");
-    writer.String(outputItemId.c_str());
-    writer.String("type");
-    writer.String("message");
-    writer.String("role");
-    writer.String("assistant");
-    writer.String("status");
-    writer.String(status.c_str());
-    writer.String("content");
-    writer.StartArray();
-    if (status != "in_progress") {
-        serializeOutputTextPart(writer, text);
-    }
-    writer.EndArray();
-    writer.EndObject();
-}
-
-void OpenAIChatCompletionsHandler::serializeOutputTextPart(Writer<StringBuffer>& writer, const std::string& text) {
-    writer.StartObject();
-    writer.String("type");
-    writer.String("output_text");
-    writer.String("text");
-    writer.String(text.c_str());
-    writer.String("annotations");
-    writer.StartArray();
-    writer.EndArray();
-    writer.EndObject();
-}
-
-std::string OpenAIChatCompletionsHandler::serializeResponsesUnaryResponse(const std::vector<ParsedOutput>& parsedOutputs,
-    ov::genai::GenerationFinishReason finishReason) const {
-    const bool isIncomplete = (finishReason == ov::genai::GenerationFinishReason::LENGTH);
-    const std::string responseStatus = isIncomplete ? "incomplete" : "completed";
-    const auto createdAt = std::chrono::duration_cast<std::chrono::seconds>(created.time_since_epoch()).count();
-    const auto completedAt = std::chrono::duration_cast<std::chrono::seconds>(std::chrono::system_clock::now().time_since_epoch()).count();
-    const std::string responseId = "resp-" + std::to_string(createdAt);
-
-    StringBuffer buffer;
-    Writer<StringBuffer> writer(buffer);
-
-    writer.StartObject();
-    writer.String("id");
-    writer.String(responseId.c_str());
-    writer.String("object");
-    writer.String("response");
-    writer.String("created_at");
-    writer.Int64(createdAt);
-    if (!isIncomplete) {
-        writer.String("completed_at");
-        writer.Int64(completedAt);
-    }
-    if (isIncomplete) {
-        writer.String("incomplete_details");
-        writer.StartObject();
-        writer.String("reason");
-        writer.String("max_tokens");
-        writer.EndObject();
-    }
-    // TODO: error not supported in unary response
-    writer.String("model");
-    writer.String(request.model.c_str());
-    writer.String("status");
-    writer.String(responseStatus.c_str());
-
-    writer.String("parallel_tool_calls");
-    writer.Bool(true);
-    // TODO: previous_response_id not supported
-    writer.String("store");
-    writer.Bool(true);
-    // TODO: temperature/top_p are only included when explicitly provided in the request
-    if (request.temperature.has_value()) {
-        writer.String("temperature");
-        writer.Double(static_cast<double>(request.temperature.value()));
-    }
-    writer.String("text");
-    writer.StartObject();
-    writer.String("format");
-    writer.StartObject();
-    writer.String("type");
-    writer.String("text");
-    writer.EndObject();
-    writer.EndObject();
-    serializeResponsesToolChoice(writer);
-    serializeResponsesTools(writer);
-    if (request.topP.has_value()) {
-        writer.String("top_p");
-        writer.Double(static_cast<double>(request.topP.value()));
-    }
-    writer.String("truncation");
-    writer.String("disabled");
-    // TODO: user not supported
-    writer.String("metadata");
-    writer.StartObject();
-    writer.EndObject();
-
-    if (request.maxTokens.has_value()) {
-        writer.String("max_output_tokens");
-        writer.Uint64(static_cast<uint64_t>(request.maxTokens.value()));
-    }
-
-    writer.String("output");
-    writer.StartArray();
-    int outputIndex = 0;
-    for (const auto& parsedOutput : parsedOutputs) {
-        // Emit reasoning output item if reasoning is available
-        if (!parsedOutput.reasoning.empty()) {
-            const std::string reasoningId = "rs-" + std::to_string(outputIndex);
-            writer.StartObject();
-            writer.String("id");
-            writer.String(reasoningId.c_str());
-            writer.String("type");
-            writer.String("reasoning");
-            writer.String("summary");
-            writer.StartArray();
-            writer.StartObject();
-            writer.String("type");
-            writer.String("summary_text");
-            writer.String("text");
-            writer.String(parsedOutput.reasoning.c_str());
-            writer.EndObject();
-            writer.EndArray();
-            writer.EndObject();
-        }
-
-        if (!parsedOutput.toolCalls.empty()) {
-            // Emit function_call output items for each tool call
-            for (const auto& toolCall : parsedOutput.toolCalls) {
-                writer.StartObject();
-                writer.String("id");
-                writer.String(toolCall.id.c_str());
-                writer.String("type");
-                writer.String("function_call");
-                writer.String("status");
-                writer.String(responseStatus.c_str());
-                writer.String("call_id");
-                writer.String(toolCall.id.c_str());
-                writer.String("name");
-                writer.String(toolCall.name.c_str());
-                writer.String("arguments");
-                writer.String(toolCall.arguments.c_str());
-                writer.EndObject();
-            }
-        }
-
-        // Always emit message output item
-        {
-            const std::string outputId = "msg-" + std::to_string(outputIndex);
-
-            writer.StartObject();
-            writer.String("id");
-            writer.String(outputId.c_str());
-            writer.String("type");
-            writer.String("message");
-            writer.String("role");
-            writer.String("assistant");
-            writer.String("status");
-            writer.String(responseStatus.c_str());
-            writer.String("content");
-            writer.StartArray();
-            serializeOutputTextPart(writer, parsedOutput.content);
-            writer.EndArray();
-            writer.EndObject();
-        }
-
-        outputIndex++;
-    }
-    writer.EndArray();
-
-    writer.String("usage");
-    writer.StartObject();
-    writer.String("input_tokens");
-    writer.Uint64(static_cast<uint64_t>(usage.promptTokens));
-    // TODO: input_tokens_details.cached_tokens not supported
-    writer.String("output_tokens");
-    writer.Uint64(static_cast<uint64_t>(usage.completionTokens));
-    // TODO: output_tokens_details.reasoning_tokens not supported
-    writer.String("total_tokens");
-    writer.Uint64(static_cast<uint64_t>(usage.calculateTotalTokens()));
-    writer.EndObject();
-
-    writer.EndObject();
-
-    return buffer.GetString();
+    return status;
 }
 
 absl::Status OpenAIChatCompletionsHandler::parseCompletionsPart() {
@@ -514,270 +119,54 @@ absl::Status OpenAIChatCompletionsHandler::parseCompletionsPart() {
     return absl::OkStatus();
 }
 
-static size_t appendChunkCallback(void* downloadedChunk, size_t size, size_t nmemb,
-    void* image) {
-    size_t realsize = size * nmemb;
-    auto& mem = *static_cast<std::string*>(image);
-    mem.append(static_cast<char*>(downloadedChunk), realsize);
-    return realsize;
-}
-#define CURL_SETOPT(setopt)   \
-    if (status == CURLE_OK) { \
-        status = setopt;      \
+absl::Status OpenAIChatCompletionsHandler::parseChatCompletionsPart(std::optional<uint32_t> maxTokensLimit, std::optional<std::string> allowedLocalMediaPath, std::optional<std::vector<std::string>> allowedMediaDomains) {
+    // messages: [{role: content}, {role: content}, ...]; required
+    auto status = parseMessages(allowedLocalMediaPath, allowedMediaDomains);
+    if (status != absl::OkStatus()) {
+        return status;
     }
-static absl::Status downloadImage(const char* url, std::string& image, const int64_t& sizeLimit) {
-    CURL* curl_handle = curl_easy_init();
-    if (!curl_handle) {
-        SPDLOG_LOGGER_ERROR(llm_calculator_logger, "Failed to initialize curl handle");
-        return absl::InternalError("Image downloading failed");
+    status = parseTools();
+    if (status != absl::OkStatus()) {
+        return status;
     }
-    auto handleGuard = std::unique_ptr<CURL, decltype(&curl_easy_cleanup)>(curl_handle, curl_easy_cleanup);
-
-    auto status = curl_easy_setopt(curl_handle, CURLOPT_URL, url);
-    CURL_SETOPT(curl_easy_setopt(curl_handle, CURLOPT_WRITEFUNCTION, appendChunkCallback))
-    CURL_SETOPT(curl_easy_setopt(curl_handle, CURLOPT_WRITEDATA, &image))
-    CURL_SETOPT(curl_easy_setopt(curl_handle, CURLOPT_SSL_OPTIONS, CURLSSLOPT_NATIVE_CA))
-    const char* envAllowRedirects = std::getenv("OVMS_MEDIA_URL_ALLOW_REDIRECTS");
-    if (envAllowRedirects != nullptr && (std::strcmp(envAllowRedirects, "1") == 0)) {
-        SPDLOG_LOGGER_TRACE(llm_calculator_logger, "URL redirects allowed");
-        CURL_SETOPT(curl_easy_setopt(curl_handle, CURLOPT_FOLLOWLOCATION, 1L))
+    // logprobs: bool; optional - defaults to false
+    auto it = doc.FindMember("logprobs");
+    if (it != doc.MemberEnd() && !it->value.IsNull()) {
+        if (!it->value.IsBool())
+            return absl::InvalidArgumentError("logprobs accepts values true or false");
+        request.logprobschat = it->value.GetBool();
     }
-    CURL_SETOPT(curl_easy_setopt(curl_handle, CURLOPT_MAXFILESIZE, sizeLimit))
-
-    if (status != CURLE_OK) {
-        SPDLOG_LOGGER_ERROR(llm_calculator_logger, "Setting curl opts failed: {}", curl_easy_strerror(status));
-        return absl::InvalidArgumentError("Image downloading failed");
+    if (request.logprobschat && request.stream) {
+        return absl::InvalidArgumentError("logprobs are not supported in streaming mode.");
     }
-
-    status = curl_easy_perform(curl_handle);
-    if (status != CURLE_OK) {
-        SPDLOG_LOGGER_ERROR(llm_calculator_logger, "Downloading image failed: {}", curl_easy_strerror(status));
-        return absl::InvalidArgumentError("Image downloading failed");
-    } else {
-        SPDLOG_LOGGER_DEBUG(llm_calculator_logger, "Downloading image succeeded, {} bytes retrieved", image.size());
-    }
-    return absl::OkStatus();
-}
-
-static bool isDomainAllowed(const std::vector<std::string>& allowedDomains, const char* url) {
-    if (allowedDomains.size() == 1 && allowedDomains[0] == "all") {
-        return true;
-    }
-    CURLUcode rc;
-    CURLU* parsedUrl = curl_url();
-    rc = curl_url_set(parsedUrl, CURLUPART_URL, url, 0);
-    if (rc) {
-        SPDLOG_LOGGER_DEBUG(llm_calculator_logger, "Parsing url {} failed", url);
-        curl_url_cleanup(parsedUrl);
-        return false;
-    }
-    char* host;
-    rc = curl_url_get(parsedUrl, CURLUPART_HOST, &host, 0);
-    if (rc) {
-        SPDLOG_LOGGER_DEBUG(llm_calculator_logger, "Parsing url {} hostname failed", url);
-        curl_url_cleanup(parsedUrl);
-        return false;
-    }
-    bool allowed = false;
-    for (const auto& allowedDomain : allowedDomains) {
-        if (allowedDomain.compare(host) == 0) {
-            allowed = true;
-            break;
+    // max_completion_tokens: uint; optional
+    it = doc.FindMember("max_completion_tokens");
+    if (it != doc.MemberEnd() && !it->value.IsNull()) {
+        if (!it->value.IsUint()) {
+            if (it->value.IsUint64())
+                return absl::InvalidArgumentError("max_completion_tokens value can't be greater than 4294967295");
+            return absl::InvalidArgumentError("max_completion_tokens is not an unsigned integer");
         }
+        if (maxTokensLimit.has_value() && it->value.GetUint() > maxTokensLimit.value())
+            return absl::InvalidArgumentError(absl::StrCat("max_completion_tokens exceeds limit provided in graph config: ", maxTokensLimit.value()));
+        request.maxTokens = it->value.GetUint();
     }
-    curl_free(host);
-    curl_url_cleanup(parsedUrl);
-    return allowed;
-}
-
-absl::Status OpenAIChatCompletionsHandler::ensureArgumentsInToolCalls(Value& messageObj, bool& jsonChanged) {
-    auto& allocator = doc.GetAllocator();
-    auto toolCallsIt = messageObj.FindMember("tool_calls");
-    if (toolCallsIt != messageObj.MemberEnd() && toolCallsIt->value.IsArray()) {
-        const auto& toolCallsArray = toolCallsIt->value.GetArray();
-        for (rapidjson::SizeType j = 0; j < toolCallsArray.Size(); ++j) {
-            auto& toolCall = toolCallsArray[j];
-            if (!toolCall.IsObject()) {
-                return absl::InvalidArgumentError("Each tool_call must be an object");
-            }
-            auto functionIt = toolCall.FindMember("function");
-            if (functionIt == toolCall.MemberEnd() || !functionIt->value.IsObject()) {
-                return absl::InvalidArgumentError("Each tool_call must have a 'function' object");
-            }
-            const auto& functionObj = functionIt->value.GetObject();
-            if (functionObj.FindMember("arguments") == functionObj.MemberEnd()) {
-                // Add "arguments": "{}"
-                rapidjson::Value argumentsKey("arguments", allocator);
-                rapidjson::Value argumentsValue;
-                argumentsValue.SetString("{}", allocator);
-                functionIt->value.GetObject().AddMember(argumentsKey, argumentsValue, allocator);
-                jsonChanged = true;
-            }
-        }
-    }
-    return absl::OkStatus();
-}
-
-absl::Status OpenAIChatCompletionsHandler::parseResponsesInput(std::optional<std::string> allowedLocalMediaPath, std::optional<std::vector<std::string>> allowedMediaDomains) {
-    auto inputIt = doc.FindMember("input");
-    if (inputIt == doc.MemberEnd()) {
-        return absl::InvalidArgumentError("input missing in request");
+    // specific part of max_tokens validation due to echo dependency
+    if (request.maxTokens == 0) {
+        return absl::InvalidArgumentError("max_tokens value should be greater than 0");
     }
 
-    if (inputIt->value.IsString()) {
-        request.prompt = inputIt->value.GetString();
-        if (request.prompt.value().empty()) {
-            return absl::InvalidArgumentError("input cannot be empty");
-        }
-
-        request.chatHistory.push_back({});
-        request.chatHistory.last()["role"] = "user";
-        request.chatHistory.last()["content"] = request.prompt.value();
-    } else if (inputIt->value.IsArray()) {
-        if (inputIt->value.GetArray().Size() == 0) {
-            return absl::InvalidArgumentError("Messages array cannot be empty");
-        }
-
-        for (size_t i = 0; i < inputIt->value.GetArray().Size(); ++i) {
-            auto& item = inputIt->value.GetArray()[i];
-            if (!item.IsObject()) {
-                return absl::InvalidArgumentError("input array items must be objects");
-            }
-
-            auto itemObj = item.GetObject();
-            auto roleIt = itemObj.FindMember("role");
-            if (roleIt == itemObj.MemberEnd() || !roleIt->value.IsString()) {
-                return absl::InvalidArgumentError("input item role is missing or invalid");
-            }
-
-            request.chatHistory.push_back({});
-            request.chatHistory.last()["role"] = roleIt->value.GetString();
-
-            auto contentIt = itemObj.FindMember("content");
-            if (contentIt == itemObj.MemberEnd()) {
-                return absl::InvalidArgumentError("input item content is missing");
-            }
-
-            if (contentIt->value.IsString()) {
-                request.chatHistory.last()["content"] = contentIt->value.GetString();
-                continue;
-            }
-
-            if (!contentIt->value.IsArray()) {
-                return absl::InvalidArgumentError("input item content must be a string or array");
-            }
-            if (contentIt->value.GetArray().Size() == 0) {
-                return absl::InvalidArgumentError("Invalid message structure - content array is empty");
-            }
-
-            std::string contentText = "";
-            for (auto& contentItem : contentIt->value.GetArray()) {
-                if (!contentItem.IsObject()) {
-                    return absl::InvalidArgumentError("input content items must be objects");
-                }
-                auto contentObj = contentItem.GetObject();
-                auto typeIt = contentObj.FindMember("type");
-                if (typeIt == contentObj.MemberEnd() || !typeIt->value.IsString()) {
-                    return absl::InvalidArgumentError("input content item type is missing or invalid");
-                }
-
-                const std::string type = typeIt->value.GetString();
-                if (type == "input_text") {
-                    auto textIt = contentObj.FindMember("text");
-                    if (textIt == contentObj.MemberEnd() || !textIt->value.IsString()) {
-                        return absl::InvalidArgumentError("input_text requires a valid text field");
-                    }
-                    contentText = textIt->value.GetString();
-                } else if (type == "input_image") {
-                    std::string imageUrl;
-                    auto imageUrlIt = contentObj.FindMember("image_url");
-                    if (imageUrlIt == contentObj.MemberEnd()) {
-                        return absl::InvalidArgumentError("input_image requires image_url field");
-                    }
-                    if (imageUrlIt->value.IsString()) {
-                        imageUrl = imageUrlIt->value.GetString();
-                    } else if (imageUrlIt->value.IsObject()) {
-                        auto imageUrlObj = imageUrlIt->value.GetObject();
-                        auto urlIt = imageUrlObj.FindMember("url");
-                        if (urlIt == imageUrlObj.MemberEnd() || !urlIt->value.IsString()) {
-                            return absl::InvalidArgumentError("input_image.image_url.url is missing or invalid");
-                        }
-                        imageUrl = urlIt->value.GetString();
-                    } else {
-                        return absl::InvalidArgumentError("input_image.image_url must be a string or object");
-                    }
-
-                    std::size_t pos = imageUrl.find(BASE64_PREFIX);
-                    std::string decoded;
-                    ov::Tensor tensor;
-                    if (pos != std::string::npos) {
-                        SPDLOG_LOGGER_TRACE(llm_calculator_logger, "Loading image from base64 string");
-                        size_t offset = pos + BASE64_PREFIX.length();
-                        if (!absl::Base64Unescape(std::string_view(imageUrl.data() + offset, imageUrl.size() - offset), &decoded)) {
-                            return absl::InvalidArgumentError("Invalid base64 string in request");
-                        }
-                        try {
-                            tensor = loadImageStbiFromMemory(decoded);
-                        } catch (std::runtime_error& e) {
-                            std::stringstream ss;
-                            ss << "Image parsing failed: " << e.what();
-                            SPDLOG_LOGGER_DEBUG(llm_calculator_logger, ss.str());
-                            return absl::InvalidArgumentError(ss.str());
-                        }
-                    } else if (std::regex_match(imageUrl.c_str(), std::regex("^(http|https|ftp|sftp|)://(.*)"))) {
-                        SPDLOG_LOGGER_TRACE(llm_calculator_logger, "Loading image using curl");
-                        if (!allowedMediaDomains.has_value() || !isDomainAllowed(allowedMediaDomains.value(), imageUrl.c_str())) {
-                            return absl::InvalidArgumentError("Given url does not match any allowed domain from allowed_media_domains");
-                        }
-                        auto status = downloadImage(imageUrl.c_str(), decoded, MAX_IMAGE_SIZE_BYTES);
-                        if (status != absl::OkStatus()) {
-                            return status;
-                        }
-                        try {
-                            tensor = loadImageStbiFromMemory(decoded);
-                        } catch (std::runtime_error& e) {
-                            std::stringstream ss;
-                            ss << "Image parsing failed: " << e.what();
-                            SPDLOG_LOGGER_DEBUG(llm_calculator_logger, ss.str());
-                            return absl::InvalidArgumentError("Image parsing failed");
-                        }
-                    } else {
-                        if (!allowedLocalMediaPath.has_value()) {
-                            return absl::InvalidArgumentError("Loading images from local filesystem is disabled.");
-                        }
-                        if (FileSystem::isPathEscaped(imageUrl)) {
-                            std::stringstream ss;
-                            ss << "Path " << imageUrl.c_str() << " escape with .. is forbidden.";
-                            SPDLOG_LOGGER_DEBUG(llm_calculator_logger, ss.str());
-                            return absl::InvalidArgumentError(ss.str());
-                        }
-                        SPDLOG_LOGGER_TRACE(llm_calculator_logger, "Loading image from local filesystem");
-                        const auto firstMissmatch = std::mismatch(imageUrl.begin(), imageUrl.end(), allowedLocalMediaPath.value().begin(), allowedLocalMediaPath.value().end());
-                        if (firstMissmatch.second != allowedLocalMediaPath.value().end()) {
-                            return absl::InvalidArgumentError("Given filepath is not subpath of allowed_local_media_path");
-                        }
-                        try {
-                            tensor = loadImageStbiFromFile(imageUrl.c_str());
-                        } catch (std::runtime_error& e) {
-                            std::stringstream ss;
-                            ss << "Image file " << imageUrl.c_str() << " parsing failed: " << e.what();
-                            SPDLOG_LOGGER_DEBUG(llm_calculator_logger, ss.str());
-                            return absl::InvalidArgumentError(ss.str());
-                        }
-                    }
-                    request.imageHistory.push_back({i, tensor});
-                } else {
-                    return absl::InvalidArgumentError("Unsupported content type. Supported types are input_text and input_image.");
-                }
-            }
-
-            request.chatHistory.last()["content"] = contentText;
-        }
-    } else {
-        return absl::InvalidArgumentError("input is not a string or array");
+    // parse response_format
+    it = doc.FindMember("response_format");
+    if (it != doc.MemberEnd()) {
+        if (it->value.IsNull())
+            return absl::OkStatus();
+        if (!it->value.IsObject())
+            return absl::InvalidArgumentError("response_format is not an object");
+        const rapidjson::Value& responseFormat = it->value;
+        request.responseFormat = convertOpenAIResponseFormatToStructuralTagStringFormat(responseFormat);
     }
 
-    SPDLOG_LOGGER_DEBUG(llm_calculator_logger, "Parsed responses input directly to chat history without mutating request JSON");
     return absl::OkStatus();
 }
 
@@ -818,7 +207,11 @@ absl::Status OpenAIChatCompletionsHandler::parseMessages(std::optional<std::stri
             }
             // Handle tool_calls array for function calling
             if (memberName == "tool_calls" && member->value.IsArray()) {
-                request.chatHistory.last()[memberName] = rapidJsonValueToJsonContainer(member->value);
+                // Use JSON string conversion for tool_calls array
+                rapidjson::StringBuffer tcBuffer;
+                rapidjson::Writer<rapidjson::StringBuffer> tcWriter(tcBuffer);
+                member->value.Accept(tcWriter);
+                request.chatHistory.last()[memberName] = ov::genai::JsonContainer::from_json_string(tcBuffer.GetString());
                 continue;
             }
             if (memberName == "content" && member->value.IsArray()) {
@@ -853,13 +246,12 @@ absl::Status OpenAIChatCompletionsHandler::parseMessages(std::optional<std::stri
                             return absl::InvalidArgumentError("Invalid message structure - image_url does not have url field");
                         }
                         std::string url = imageUrl["url"].GetString();
-                        std::string pattern = "base64,";
-                        std::size_t pos = url.find(pattern);
+                        std::size_t pos = url.find(BASE64_PREFIX);
                         std::string decoded;
                         ov::Tensor tensor;
                         if (pos != std::string::npos) {
                             SPDLOG_LOGGER_TRACE(llm_calculator_logger, "Loading image from base64 string");
-                            size_t offset = pos + pattern.length();
+                            size_t offset = pos + BASE64_PREFIX.length();
                             if (!absl::Base64Unescape(std::string_view(url.data() + offset, url.size() - offset), &decoded)) {
                                 return absl::InvalidArgumentError("Invalid base64 string in request");
                             }
@@ -873,11 +265,10 @@ absl::Status OpenAIChatCompletionsHandler::parseMessages(std::optional<std::stri
                             }
                         } else if (std::regex_match(url.c_str(), std::regex("^(http|https|ftp|sftp|)://(.*)"))) {
                             SPDLOG_LOGGER_TRACE(llm_calculator_logger, "Loading image using curl");
-                            int64_t sizeLimit = 20000000;  // restrict single image size to 20MB
                             if (!allowedMediaDomains.has_value() || !isDomainAllowed(allowedMediaDomains.value(), url.c_str())) {
                                 return absl::InvalidArgumentError("Given url does not match any allowed domain from allowed_media_domains");
                             }
-                            auto status = downloadImage(url.c_str(), decoded, sizeLimit);
+                            auto status = downloadImage(url.c_str(), decoded, MAX_IMAGE_SIZE_BYTES);
                             if (status != absl::OkStatus()) {
                                 return status;
                             }
@@ -889,42 +280,6 @@ absl::Status OpenAIChatCompletionsHandler::parseMessages(std::optional<std::stri
                                 SPDLOG_LOGGER_DEBUG(llm_calculator_logger, ss.str());
                                 return absl::InvalidArgumentError("Image parsing failed");
                             }
-                            std::string url = imageUrl["url"].GetString();
-                            std::size_t pos = url.find(BASE64_PREFIX);
-                            std::string decoded;
-                            ov::Tensor tensor;
-                            if (pos != std::string::npos) {
-                                SPDLOG_LOGGER_TRACE(llm_calculator_logger, "Loading image from base64 string");
-                                size_t offset = pos + BASE64_PREFIX.length();
-                                if (!absl::Base64Unescape(std::string_view(url.data() + offset, url.size() - offset), &decoded)) {
-                                    return absl::InvalidArgumentError("Invalid base64 string in request");
-                                }
-                                try {
-                                    tensor = loadImageStbiFromMemory(decoded);
-                                } catch (std::runtime_error& e) {
-                                    std::stringstream ss;
-                                    ss << "Image parsing failed: " << e.what();
-                                    SPDLOG_LOGGER_DEBUG(llm_calculator_logger, ss.str());
-                                    return absl::InvalidArgumentError(ss.str());
-                                }
-                            } else if (std::regex_match(url.c_str(), std::regex("^(http|https|ftp|sftp|)://(.*)"))) {
-                                SPDLOG_LOGGER_TRACE(llm_calculator_logger, "Loading image using curl");
-                                if (!allowedMediaDomains.has_value() || !isDomainAllowed(allowedMediaDomains.value(), url.c_str())) {
-                                    return absl::InvalidArgumentError("Given url does not match any allowed domain from allowed_media_domains");
-                                }
-                                auto status = downloadImage(url.c_str(), decoded, MAX_IMAGE_SIZE_BYTES);
-                                if (status != absl::OkStatus()) {
-                                    return status;
-                                }
-                                try {
-                                    tensor = loadImageStbiFromMemory(decoded);
-                                } catch (std::runtime_error& e) {
-                                    std::stringstream ss;
-                                    ss << "Image parsing failed: " << e.what();
-                                    SPDLOG_LOGGER_DEBUG(llm_calculator_logger, ss.str());
-                                    return absl::InvalidArgumentError("Image parsing failed");
-                                }
-
                         } else {
                             if (!allowedLocalMediaPath.has_value()) {
                                 return absl::InvalidArgumentError("Loading images from local filesystem is disabled.");
@@ -989,793 +344,10 @@ absl::Status OpenAIChatCompletionsHandler::parseMessages(std::optional<std::stri
     return absl::OkStatus();
 }
 
-absl::Status OpenAIChatCompletionsHandler::parseTools() {
-    auto toolChoiceIt = doc.FindMember("tool_choice");
-    std::string toolChoice{"auto"};
-    if (toolChoiceIt != doc.MemberEnd() && !toolChoiceIt->value.IsNull()) {
-        if (toolChoiceIt->value.IsString()) {
-            toolChoice = toolChoiceIt->value.GetString();
-            if (toolChoice != "none" && toolChoice != "auto" && toolChoice != "required")
-                return absl::InvalidArgumentError("tool_choice should be either 'none' or 'auto' or 'required'");
-        } else if (toolChoiceIt->value.IsObject()) {
-            auto toolChoiceObj = toolChoiceIt->value.GetObject();
-            auto toolChoiceFunctionIt = toolChoiceObj.FindMember("function");
-            if (toolChoiceFunctionIt != toolChoiceObj.MemberEnd() && toolChoiceFunctionIt->value.IsObject()) {
-                auto nameIt = toolChoiceFunctionIt->value.GetObject().FindMember("name");
-                if (nameIt != toolChoiceFunctionIt->value.GetObject().MemberEnd() && nameIt->value.IsString()) {
-                    toolChoice = nameIt->value.GetString();
-                } else {
-                    return absl::InvalidArgumentError("tool_choice.function.name is not a valid string");
-                }
-            } else {
-                auto typeIt = toolChoiceObj.FindMember("type");
-                auto nameIt = toolChoiceObj.FindMember("name");
-                if (typeIt != toolChoiceObj.MemberEnd() && typeIt->value.IsString() && std::string(typeIt->value.GetString()) == "function") {
-                    if (nameIt == toolChoiceObj.MemberEnd() || !nameIt->value.IsString()) {
-                        return absl::InvalidArgumentError("tool_choice.name is not a valid string");
-                    }
-                    toolChoice = nameIt->value.GetString();
-                } else {
-                    return absl::InvalidArgumentError("tool_choice.function is not a valid JSON object");
-                }
-            }
-        } else {
-            return absl::InvalidArgumentError("tool_choice is not a valid JSON object or string");
-        }
-    }
-    bool jsonChanged = false;
-    if (toolChoice == "none") {
-        // remove tools from the request
-        doc.RemoveMember("tools");
-        jsonChanged = true;
-    }
-    auto it = doc.FindMember("tools");
-    if (it != doc.MemberEnd() && !it->value.IsNull()) {
-        if (!it->value.IsArray())
-            return absl::InvalidArgumentError("Tools are not an array");
-        for (size_t i = 0; i < it->value.GetArray().Size();) {
-            auto& obj = it->value.GetArray()[i];
-            if (!obj.IsObject())
-                return absl::InvalidArgumentError("Tool is not a JSON object");
-            rapidjson::Value* parametersValue = nullptr;
-            std::string functionName;
-
-            auto functionIt = obj.FindMember("function");
-            if (functionIt != obj.MemberEnd()) {
-                if (!functionIt->value.IsObject()) {
-                    return absl::InvalidArgumentError("Function is not a valid JSON object");
-                }
-                auto& functionObj = functionIt->value;
-                auto nameIt = functionObj.GetObject().FindMember("name");
-                if (nameIt == functionObj.GetObject().MemberEnd() || !nameIt->value.IsString()) {
-                    return absl::InvalidArgumentError("Function object does not contain a valid name field");
-                }
-                functionName = nameIt->value.GetString();
-                auto parametersIt = functionObj.GetObject().FindMember("parameters");
-                if (parametersIt != functionObj.GetObject().MemberEnd()) {
-                    parametersValue = &parametersIt->value;
-                }
-            } else {
-                auto typeIt = obj.FindMember("type");
-                if (typeIt == obj.MemberEnd() || !typeIt->value.IsString()) {
-                    return absl::InvalidArgumentError("Tool type is missing or invalid");
-                }
-                if (std::string(typeIt->value.GetString()) != "function") {
-                    return absl::InvalidArgumentError("Only function tools are supported");
-                }
-
-                auto nameIt = obj.FindMember("name");
-                if (nameIt == obj.MemberEnd() || !nameIt->value.IsString()) {
-                    return absl::InvalidArgumentError("Function object does not contain a valid name field");
-                }
-                functionName = nameIt->value.GetString();
-
-                auto parametersIt = obj.FindMember("parameters");
-                if (parametersIt != obj.MemberEnd()) {
-                    parametersValue = &parametersIt->value;
-                }
-            }
-
-            // If toolChoice is set to "auto", we keep all tools
-            // If toolChoice is set to a specific function name, we keep only that tool
-            if (toolChoice != "auto" && toolChoice != "required" && toolChoice != functionName) {
-                it->value.Erase(&obj);
-                jsonChanged = true;
-                continue;
-            }
-
-            i++;
-            // If we keep the tool, add tool name and schema to the request
-            if (parametersValue != nullptr) {
-                if (!parametersValue->IsObject()) {
-                    return absl::InvalidArgumentError("Function parameters are not a valid JSON object");
-                }
-                // Dump parameters object to string since this is the schema format expected by GenAI
-                // Keep the rapidjson::Value pointer as well to avoid re-parsing in outputParsers
-                rapidjson::StringBuffer buffer;
-                rapidjson::Writer<rapidjson::StringBuffer> writer(buffer);
-                parametersValue->Accept(writer);
-                std::string parametersStr = buffer.GetString();
-                ToolSchemaWrapper schemaReprs{parametersValue, std::move(parametersStr)};
-                request.toolNameSchemaMap[functionName] = std::move(schemaReprs);
-            }
-        }
-    } else {
-        toolChoice = "none";  // If tools are not provided, set toolChoice to "none"
-    }
-
-    request.toolChoice = toolChoice;
-    if (jsonChanged) {
-        StringBuffer buffer;
-        Writer<StringBuffer> writer(buffer);
-        doc.Accept(writer);
-        request.processedJson = buffer.GetString();
-    }
-    return absl::OkStatus();
-}
-
-absl::StatusOr<std::optional<ov::genai::JsonContainer>> OpenAIChatCompletionsHandler::parseToolsToJsonContainer() {
-    auto it = doc.FindMember("tools");
-    if (it == doc.MemberEnd() || it->value.IsNull()) {
-        return std::nullopt;
-    }
-    try {
-        return rapidJsonValueToJsonContainer(it->value);
-    } catch (const std::exception& e) {
-        SPDLOG_LOGGER_DEBUG(llm_calculator_logger, "Direct tools conversion to JsonContainer failed: {}. Falling back to JSON string conversion.", e.what());
-        try {
-            rapidjson::StringBuffer toolsBuffer;
-            rapidjson::Writer<rapidjson::StringBuffer> toolsWriter(toolsBuffer);
-            it->value.Accept(toolsWriter);
-            return ov::genai::JsonContainer::from_json_string(toolsBuffer.GetString());
-        } catch (const std::exception& fallbackEx) {
-            SPDLOG_LOGGER_DEBUG(llm_calculator_logger, "Fallback tools conversion failed: {}", fallbackEx.what());
-            return absl::InvalidArgumentError(absl::StrCat("Invalid tools payload: ", fallbackEx.what()));
-        }
-    }
-}
-
-absl::StatusOr<std::optional<ov::genai::JsonContainer>> OpenAIChatCompletionsHandler::parseChatTemplateKwargsToJsonContainer() {
-    auto it = doc.FindMember("chat_template_kwargs");
-    if (it == doc.MemberEnd() || it->value.IsNull()) {
-        return std::nullopt;
-    }
-    if (!it->value.IsObject()) {
-        return absl::InvalidArgumentError("chat_template_kwargs must be an object");
-    }
-    try {
-        return rapidJsonValueToJsonContainer(it->value);
-    } catch (const std::exception& e) {
-        SPDLOG_LOGGER_DEBUG(llm_calculator_logger, "Direct chat_template_kwargs conversion to JsonContainer failed: {}. Falling back to JSON string conversion.", e.what());
-        try {
-            rapidjson::StringBuffer buffer;
-            rapidjson::Writer<rapidjson::StringBuffer> writer(buffer);
-            it->value.Accept(writer);
-            return ov::genai::JsonContainer::from_json_string(buffer.GetString());
-        } catch (const std::exception& fallbackEx) {
-            SPDLOG_LOGGER_DEBUG(llm_calculator_logger, "Fallback chat_template_kwargs conversion failed: {}", fallbackEx.what());
-            return absl::InvalidArgumentError(absl::StrCat("Invalid chat_template_kwargs payload: ", fallbackEx.what()));
-        }
-    }
-}
-
-const bool OpenAIChatCompletionsHandler::areToolsAvailable() const {
-    return !request.toolNameSchemaMap.empty();
-}
-
-const OpenAIChatCompletionsRequest& OpenAIChatCompletionsHandler::getRequest() const {
-    return request;
-}
-
-const std::string& OpenAIChatCompletionsHandler::getProcessedJson() const {
-    return request.processedJson;
-}
-const ImageHistory& OpenAIChatCompletionsHandler::getImageHistory() const {
-    return request.imageHistory;
-}
-
-ov::genai::ChatHistory& OpenAIChatCompletionsHandler::getChatHistory() {
-    return request.chatHistory;
-}
-
-std::optional<int> OpenAIChatCompletionsHandler::getMaxTokens() const {
-    return request.maxTokens;
-}
-
-std::optional<std::string> OpenAIChatCompletionsHandler::getResponseFormat() const {
-    return request.responseFormat;
-}
-
-std::string convertOpenAIResponseFormatToStructuralTagStringFormat(const rapidjson::Value& openAIFormat) {
-    // Build the new object: {"type": "structural_tag", "format": <openAIFormat>}
-    // If response_format has {"json_schema": {"schema": {...}}}, flatten it to {"json_schema": {...}}
-    rapidjson::Document flatFormatDoc;
-    flatFormatDoc.CopyFrom(openAIFormat, flatFormatDoc.GetAllocator());
-
-    if (flatFormatDoc.HasMember("json_schema") && flatFormatDoc["json_schema"].IsObject()) {
-        auto& jsonSchema = flatFormatDoc["json_schema"];
-        if (jsonSchema.HasMember("schema") && jsonSchema["schema"].IsObject()) {
-            // Move all members from "schema" to "json_schema"
-            rapidjson::Value schemaObjCopy;
-            schemaObjCopy.CopyFrom(jsonSchema["schema"], flatFormatDoc.GetAllocator());  // Make a copy as we will modify jsonSchema
-            for (auto itr = schemaObjCopy.MemberBegin(); itr != schemaObjCopy.MemberEnd(); ++itr) {
-                rapidjson::Value key;
-                key.CopyFrom(itr->name, flatFormatDoc.GetAllocator());
-                rapidjson::Value value;
-                value.CopyFrom(itr->value, flatFormatDoc.GetAllocator());
-                jsonSchema.AddMember(key, value, flatFormatDoc.GetAllocator());
-            }
-            // Remove the "schema" member
-            jsonSchema.RemoveMember("schema");
-        }
-    }
-
-    // Serialize the flattened response_format object
-    rapidjson::StringBuffer formatBuffer;
-    rapidjson::Writer<rapidjson::StringBuffer> formatWriter(formatBuffer);
-    flatFormatDoc.Accept(formatWriter);
-
-    // Build the new object: {"type": "structural_tag", "format": <flattened>}
-    rapidjson::StringBuffer buffer;
-    rapidjson::Writer<rapidjson::StringBuffer> writer(buffer);
-    writer.StartObject();
-    writer.Key("type");
-    writer.String("structural_tag");
-    writer.Key("format");
-    writer.RawValue(formatBuffer.GetString(), formatBuffer.GetSize(), rapidjson::kObjectType);
-    writer.EndObject();
-    return buffer.GetString();
-}
-
-absl::Status OpenAIChatCompletionsHandler::parseChatCompletionsPart(std::optional<uint32_t> maxTokensLimit, std::optional<std::string> allowedLocalMediaPath, std::optional<std::vector<std::string>> allowedMediaDomains) {
-    // messages: [{role: content}, {role: content}, ...]; required
-    auto status = parseMessages(allowedLocalMediaPath, allowedMediaDomains);
-    if (status != absl::OkStatus()) {
-        return status;
-    }
-    status = parseTools();
-    if (status != absl::OkStatus()) {
-        return status;
-    }
-    // logprobs: bool; optional - defaults to false
-    auto it = doc.FindMember("logprobs");
-    if (it != doc.MemberEnd() && !it->value.IsNull()) {
-        if (!it->value.IsBool())
-            return absl::InvalidArgumentError("logprobs accepts values true or false");
-        request.logprobschat = it->value.GetBool();
-    }
-    if (request.logprobschat && request.stream) {
-        return absl::InvalidArgumentError("logprobs are not supported in streaming mode.");
-    }
-    // max_completion_tokens: uint; optional
-    it = doc.FindMember("max_completion_tokens");
-    if (it != doc.MemberEnd() && !it->value.IsNull()) {
-        if (!it->value.IsUint()) {
-            if (it->value.IsUint64())
-                return absl::InvalidArgumentError("max_completion_tokens value can't be greater than 4294967295");
-            return absl::InvalidArgumentError("max_completion_tokens is not an unsigned integer");
-        }
-        if (maxTokensLimit.has_value() && it->value.GetUint() > maxTokensLimit.value())
-            return absl::InvalidArgumentError(absl::StrCat("max_completion_tokens exceeds limit provided in graph config: ", maxTokensLimit.value()));
-        request.maxTokens = it->value.GetUint();
-    }
-    // specific part of max_tokens validation due to echo dependency
-    if (request.maxTokens == 0) {
-        return absl::InvalidArgumentError("max_tokens value should be greater than 0");
-    }
-
-    // parse response_format
-    it = doc.FindMember("response_format");
-    if (it != doc.MemberEnd()) {
-        if (it->value.IsNull())
-            return absl::OkStatus();
-        if (!it->value.IsObject())
-            return absl::InvalidArgumentError("response_format is not an object");
-        const rapidjson::Value& responseFormat = it->value;
-        request.responseFormat = convertOpenAIResponseFormatToStructuralTagStringFormat(responseFormat);
-    }
-
-    return absl::OkStatus();
-}
-
-absl::Status OpenAIChatCompletionsHandler::parseResponsesPart(std::optional<uint32_t> maxTokensLimit, std::optional<std::string> allowedLocalMediaPath, std::optional<std::vector<std::string>> allowedMediaDomains) {
-    // input: string; required
-    auto it = doc.FindMember("input");
-    if (it == doc.MemberEnd()) {
-        return absl::InvalidArgumentError("input missing in request");
-    }
-
-    auto messagesStatus = parseResponsesInput(allowedLocalMediaPath, allowedMediaDomains);
-    if (!messagesStatus.ok()) {
-        return messagesStatus;
-    }
-
-    // reasoning: object; optional
-    // OpenAI Responses API reasoning parameter. Any effort value enables thinking mode.
-    it = doc.FindMember("reasoning");
-    if (it != doc.MemberEnd() && !it->value.IsNull()) {
-        if (!it->value.IsObject()) {
-            return absl::InvalidArgumentError("reasoning is not an object");
-        }
-        const auto& reasoningObj = it->value;
-        auto effortIt = reasoningObj.FindMember("effort");
-        if (effortIt != reasoningObj.MemberEnd() && !effortIt->value.IsNull()) {
-            if (!effortIt->value.IsString()) {
-                return absl::InvalidArgumentError("reasoning.effort is not a string");
-            }
-            const std::string effort = effortIt->value.GetString();
-            if (effort != "low" && effort != "medium" && effort != "high") {
-                return absl::InvalidArgumentError("reasoning.effort must be one of: low, medium, high");
-            }
-            // Inject enable_thinking: true into chat_template_kwargs if not already explicitly set
-            auto kwargsIt = doc.FindMember("chat_template_kwargs");
-            if (kwargsIt == doc.MemberEnd()) {
-                rapidjson::Value kwargs(rapidjson::kObjectType);
-                kwargs.AddMember("enable_thinking", true, doc.GetAllocator());
-                doc.AddMember("chat_template_kwargs", kwargs, doc.GetAllocator());
-            } else if (kwargsIt->value.IsObject()) {
-                auto enableThinkingIt = kwargsIt->value.FindMember("enable_thinking");
-                if (enableThinkingIt == kwargsIt->value.MemberEnd()) {
-                    kwargsIt->value.AddMember("enable_thinking", true, doc.GetAllocator());
-                }
-                // If enable_thinking is already set explicitly, the user's value takes precedence
-            }
-        }
-        // summary field is accepted but ignored
-    }
-
-#if (PYTHON_DISABLE == 0)
-    // Build processedJson with "messages" array from chatHistory so that
-    // the Python chat template path (which reads request_json["messages"])
-    // can consume Responses API input without a separate code path.
-    {
-        Document processedDoc;
-        processedDoc.SetObject();
-        auto& alloc = processedDoc.GetAllocator();
-
-        Value messagesArray(kArrayType);
-        for (size_t i = 0; i < request.chatHistory.size(); ++i) {
-            Value msgObj(kObjectType);
-            auto role = request.chatHistory[i]["role"].as_string();
-            if (role.has_value()) {
-                msgObj.AddMember("role", Value(role.value().c_str(), alloc), alloc);
-            }
-            auto content = request.chatHistory[i]["content"].as_string();
-            if (content.has_value()) {
-                msgObj.AddMember("content", Value(content.value().c_str(), alloc), alloc);
-            }
-            messagesArray.PushBack(msgObj, alloc);
-        }
-        processedDoc.AddMember("messages", messagesArray, alloc);
-
-        // Copy tools from original doc if present
-        auto toolsIt = doc.FindMember("tools");
-        if (toolsIt != doc.MemberEnd() && !toolsIt->value.IsNull()) {
-            Value toolsCopy(toolsIt->value, alloc);
-            processedDoc.AddMember("tools", toolsCopy, alloc);
-        }
-
-        // Copy chat_template_kwargs from original doc if present
-        auto kwargsIt = doc.FindMember("chat_template_kwargs");
-        if (kwargsIt != doc.MemberEnd() && !kwargsIt->value.IsNull()) {
-            Value kwargsCopy(kwargsIt->value, alloc);
-            processedDoc.AddMember("chat_template_kwargs", kwargsCopy, alloc);
-        }
-
-        StringBuffer buffer;
-        Writer<StringBuffer> writer(buffer);
-        processedDoc.Accept(writer);
-        request.processedJson = buffer.GetString();
-    }
-#endif
-    // logprobs: bool; optional - defaults to false
-    it = doc.FindMember("logprobs");
-    if (it != doc.MemberEnd() && !it->value.IsNull()) {
-        if (!it->value.IsBool())
-            return absl::InvalidArgumentError("logprobs accepts values true or false");
-        request.logprobschat = it->value.GetBool();
-    }
-    if (request.logprobschat && request.stream) {
-        return absl::InvalidArgumentError("logprobs are not supported in streaming mode.");
-    }
-
-    auto toolsStatus = parseTools();
-    if (!toolsStatus.ok()) {
-        return toolsStatus;
-    }
-
-    // max_output_tokens: uint; optional
-    // OpenAI Responses API uses this field for output token limit.
-    it = doc.FindMember("max_output_tokens");
-    if (it != doc.MemberEnd() && !it->value.IsNull()) {
-        if (!it->value.IsUint()) {
-            if (it->value.IsUint64())
-                return absl::InvalidArgumentError("max_output_tokens value can't be greater than 4294967295");
-            return absl::InvalidArgumentError("max_output_tokens is not an unsigned integer");
-        }
-        if (maxTokensLimit.has_value() && it->value.GetUint() > maxTokensLimit.value())
-            return absl::InvalidArgumentError(absl::StrCat("max_output_tokens exceeds limit provided in graph config: ", maxTokensLimit.value()));
-        request.maxTokens = it->value.GetUint();
-    }
-
-    // specific part of max_output_tokens validation
-    if (request.maxTokens == 0) {
-        return absl::InvalidArgumentError("max_output_tokens value should be greater than 0");
-    }
-
-    // parse response_format
-    it = doc.FindMember("response_format");
-    if (it != doc.MemberEnd()) {
-        if (it->value.IsNull())
-            return absl::OkStatus();
-        if (!it->value.IsObject())
-            return absl::InvalidArgumentError("response_format is not an object");
-        const rapidjson::Value& responseFormat = it->value;
-        request.responseFormat = convertOpenAIResponseFormatToStructuralTagStringFormat(responseFormat);
-    }
-
-    return absl::OkStatus();
-}
-
-absl::Status OpenAIChatCompletionsHandler::parseCommonPart(std::optional<uint32_t> maxTokensLimit, uint32_t bestOfLimit, std::optional<uint32_t> maxModelLength) {
-    OVMS_PROFILE_FUNCTION();
-    // stream: bool; optional
-    if (!doc.IsObject())
-        return absl::InvalidArgumentError("Received json is not an object");
-    auto it = doc.FindMember("stream");
-    if (it != doc.MemberEnd() && !it->value.IsNull()) {
-        if (!it->value.IsBool())
-            return absl::InvalidArgumentError("Stream is not bool");
-        request.stream = it->value.GetBool();
-    }
-
-    it = doc.FindMember("stream_options");
-    if (it != doc.MemberEnd() && !it->value.IsNull()) {
-        if (!request.stream)
-            return absl::InvalidArgumentError("stream_options provided, but stream not set to true");
-        if (!it->value.IsObject())
-            return absl::InvalidArgumentError("stream_options is not an object");
-        auto streamOptionsObj = it->value.GetObject();
-
-        size_t streamOptionsFound = 0;
-        it = streamOptionsObj.FindMember("include_usage");
-        if (it != streamOptionsObj.MemberEnd()) {
-            if (!it->value.IsBool())
-                return absl::InvalidArgumentError("stream_options.include_usage is not a boolean");
-            request.streamOptions.includeUsage = it->value.GetBool();
-            streamOptionsFound++;
-        }
-
-        if (streamOptionsObj.MemberCount() > streamOptionsFound) {
-            return absl::InvalidArgumentError("Found unexpected stream options. Properties accepted in stream_options: include_usage");
-        }
-    }
-
-    // model: string; required
-    it = doc.FindMember("model");
-    if (it != doc.MemberEnd()) {
-        if (!it->value.IsString())
-            return absl::InvalidArgumentError("model is not a string");
-        request.model = it->value.GetString();
-    } else {
-        return absl::InvalidArgumentError("model missing in request");
-    }
-
-    // ignore_eos: bool; optional - defaults to false
-    // Extension, unsupported by OpenAI API, however supported by vLLM and CB lib
-    it = doc.FindMember("ignore_eos");
-    if (it != doc.MemberEnd() && !it->value.IsNull()) {
-        if (!it->value.IsBool())
-            return absl::InvalidArgumentError("ignore_eos accepts values true or false");
-        request.ignoreEOS = it->value.GetBool();
-    }
-
-    // max_tokens: uint; optional
-    // Common part checked here, specific parts are checked in parseCompletionsPart and parseChatCompletionsPart
-    // TODO: Deprecated - this will need to be removed in the future
-    // Not applicable for RESPONSES endpoint which uses max_output_tokens instead
-    if (endpoint != Endpoint::RESPONSES) {
-        it = doc.FindMember("max_tokens");
-        if (it != doc.MemberEnd()) {
-            if (!it->value.IsUint()) {
-                if (it->value.IsUint64())
-                    return absl::InvalidArgumentError("max_tokens value can't be greater than 4294967295");
-                return absl::InvalidArgumentError("max_tokens is not an unsigned integer");
-            }
-            if (maxTokensLimit.has_value() && !(it->value.GetUint() < maxTokensLimit.value()))
-                return absl::InvalidArgumentError(absl::StrCat("max_tokens exceeds limit provided in graph config: ", maxTokensLimit.value()));
-            request.maxTokens = it->value.GetUint();
-        } else {
-            if (maxTokensLimit.has_value()) {
-                request.maxTokens = maxTokensLimit.value();
-            }
-        }
-    } else {
-        if (maxTokensLimit.has_value()) {
-            request.maxTokens = maxTokensLimit.value();
-        }
-    }
-
-    // frequency_penalty: float; optional - defaults to 0
-    it = doc.FindMember("frequency_penalty");
-    if (it != doc.MemberEnd() && !it->value.IsNull()) {
-        if (!it->value.IsDouble() && !it->value.IsInt())
-            return absl::InvalidArgumentError("frequency_penalty is not a valid number");
-        request.frequencyPenalty = it->value.GetDouble();
-        if (request.frequencyPenalty < -2.0f || request.frequencyPenalty > 2.0f)
-            return absl::InvalidArgumentError("frequency_penalty out of range(-2.0, 2.0)");
-    }
-
-    // presence_penalty: float; optional - defaults to 0
-    it = doc.FindMember("presence_penalty");
-    if (it != doc.MemberEnd() && !it->value.IsNull()) {
-        if (!it->value.IsDouble() && !it->value.IsInt())
-            return absl::InvalidArgumentError("presence_penalty is not a valid number");
-        request.presencePenalty = it->value.GetDouble();
-        if (request.presencePenalty < -2.0f || request.presencePenalty > 2.0f)
-            return absl::InvalidArgumentError("presence_penalty out of range(-2.0, 2.0)");
-    }
-
-    // repetition_penalty: float; optional - defaults to 1.0
-    // Extension, unsupported by OpenAI API, however supported by vLLM and CB lib
-    it = doc.FindMember("repetition_penalty");
-    if (it != doc.MemberEnd() && !it->value.IsNull()) {
-        if (!it->value.IsDouble() && !it->value.IsInt())
-            return absl::InvalidArgumentError("repetition_penalty is not a valid number");
-        request.repetitionPenalty = it->value.GetDouble();
-    }
-
-    // length_penalty: float; optional - defaults to 1.0
-    // Extension, unsupported by OpenAI API however supported by vLLM and CB lib
-    it = doc.FindMember("length_penalty");
-    if (it != doc.MemberEnd() && !it->value.IsNull()) {
-        if (!it->value.IsDouble() && !it->value.IsInt())
-            return absl::InvalidArgumentError("length_penalty is not a valid number");
-        request.lengthPenalty = it->value.GetDouble();
-    }
-
-    // temperature: float; optional - defaults to 1.0
-    it = doc.FindMember("temperature");
-    if (it != doc.MemberEnd() && !it->value.IsNull()) {
-        if (!it->value.IsDouble() && !it->value.IsInt())
-            return absl::InvalidArgumentError("temperature is not a valid number");
-        request.temperature = it->value.GetDouble();
-        if (request.temperature < 0.0f || request.temperature > 2.0f)
-            return absl::InvalidArgumentError("temperature out of range(0.0, 2.0)");
-    }
-
-    // top_p: float; optional - defaults to 1
-    it = doc.FindMember("top_p");
-    if (it != doc.MemberEnd() && !it->value.IsNull()) {
-        if (!it->value.IsDouble() && !it->value.IsInt())
-            return absl::InvalidArgumentError("top_p is not a valid number");
-        request.topP = it->value.GetDouble();
-        if (request.topP < 0.0f || request.topP > 1.0f)
-            return absl::InvalidArgumentError("top_p out of range(0.0, 1.0)");
-    }
-
-    // top_k: int; optional - defaults to 0
-    // Extension, unsupported by OpenAI API, however supported by vLLM and CB lib
-    it = doc.FindMember("top_k");
-    if (it != doc.MemberEnd() && !it->value.IsNull()) {
-        if (!it->value.IsInt())
-            return absl::InvalidArgumentError("top_k is not an integer");
-        request.topK = it->value.GetInt();
-    }
-
-    // seed: int; optional - defaults to 0 (not set)
-    it = doc.FindMember("seed");
-    if (it != doc.MemberEnd() && !it->value.IsNull()) {
-        if (!it->value.IsUint())
-            return absl::InvalidArgumentError("seed is not an unsigned integer");
-        request.seed = it->value.GetUint();
-    }
-
-    // stop: string or array; optional - defaults to null (not set)
-    it = doc.FindMember("stop");
-    if (it != doc.MemberEnd() && !it->value.IsNull()) {
-        if (it->value.IsString()) {
-            request.stop = std::set<std::string>{it->value.GetString()};
-        } else if (it->value.IsArray()) {
-            auto stopArray = it->value.GetArray();
-            if (stopArray.Size() > DEFAULT_MAX_STOP_WORDS) {
-                std::stringstream ss;
-                ss << "stop array must have no more than " << DEFAULT_MAX_STOP_WORDS << " strings";
-                return absl::InvalidArgumentError(ss.str());
-            }
-            if (!stopArray.Empty()) {
-                request.stop = std::set<std::string>{};
-                for (size_t i = 0; i < stopArray.Size(); i++) {
-                    const auto& element = stopArray[i];
-                    if (!element.IsString())
-                        return absl::InvalidArgumentError("stop array contains non string element");
-                    request.stop->insert(element.GetString());
-                }
-            }
-        } else {
-            return absl::InvalidArgumentError("stop is not a string or array of strings");
-        }
-    }
-
-    // include_stop_str_in_output: bool; optional - defaults to false
-    // Extension, unsupported by OpenAI API, however supported by vLLM and CB lib
-
-    // If stream is true, then include stop string in output by default
-    if (request.stream) {
-        request.includeStopStrInOutput = true;
-    }
-
-    it = doc.FindMember("include_stop_str_in_output");
-    if (it != doc.MemberEnd() && !it->value.IsNull()) {
-        if (!it->value.IsBool())
-            return absl::InvalidArgumentError("include_stop_str_in_output accepts values true or false");
-        if (!it->value.GetBool() && request.stream)
-            return absl::InvalidArgumentError("include_stop_str_in_output cannot be set to false if streaming is used");
-        request.includeStopStrInOutput = it->value.GetBool();
-    }
-
-    // best_of: int; optional - defaults to 1
-    // Extension, unsupported by OpenAI API, however supported by vLLM, supported in CB lib by mapping to group_size param
-    // Not supported for RESPONSES streaming - output_index is hardcoded to 0
-    it = doc.FindMember("best_of");
-    if (it != doc.MemberEnd() && !it->value.IsNull()) {
-        if (!it->value.IsUint())
-            return absl::InvalidArgumentError("best_of is not an unsigned integer");
-        if (it->value.GetUint() == 0)
-            return absl::InvalidArgumentError("best_of value should be greater than 0");
-        if (!(it->value.GetUint() < bestOfLimit))
-            return absl::InvalidArgumentError(absl::StrCat("best_of exceeds limit provided in graph config: ", bestOfLimit));
-        if (request.stream)
-            return absl::InvalidArgumentError("best_of cannot be used in streaming mode");
-        request.bestOf = it->value.GetUint();
-    }
-
-    // n: int; optional - defaults to 1
-    // Supported by OpenAI API and vLLM, supported in CB lib by mapping to num_return_sequences param
-    // Not supported for RESPONSES streaming - output_index is hardcoded to 0
-    it = doc.FindMember("n");
-    if (it != doc.MemberEnd() && !it->value.IsNull()) {
-        if (!it->value.IsUint())
-            return absl::InvalidArgumentError("n is not an unsigned integer");
-        if (it->value.GetUint() == 0)
-            return absl::InvalidArgumentError("n value should be greater than 0");
-        if (endpoint == Endpoint::RESPONSES && request.stream && it->value.GetUint() > 1)
-            return absl::InvalidArgumentError("n greater than 1 is not supported for Responses API streaming");
-        size_t bestOf = request.bestOf.has_value() ? request.bestOf.value() : 1;  // 1 is default best_of value
-        if (bestOf < it->value.GetUint()) {
-            return absl::InvalidArgumentError("n value cannot be greater than best_of");
-        }
-        request.numReturnSequences = it->value.GetUint();
-    }
-
-    // Assisted decoding specific parameters
-
-    auto numAssistantTokensIt = doc.FindMember("num_assistant_tokens");
-    auto assistantConfidenceThresholdIt = doc.FindMember("assistant_confidence_threshold");
-    auto maxNgramSizeIt = doc.FindMember("max_ngram_size");
-
-    bool numAssistantTokensItHasValue = (numAssistantTokensIt != doc.MemberEnd() && !numAssistantTokensIt->value.IsNull());
-    bool assistantConfidenceThresholdItHasValue = (assistantConfidenceThresholdIt != doc.MemberEnd() && !assistantConfidenceThresholdIt->value.IsNull());
-    bool maxNgramSizeItHasValue = (maxNgramSizeIt != doc.MemberEnd() && !maxNgramSizeIt->value.IsNull());
-
-    if (numAssistantTokensItHasValue) {
-        request.numAssistantTokens = numAssistantTokensIt->value.GetUint();
-    }
-    if (assistantConfidenceThresholdItHasValue) {
-        request.assistantConfidenceThreshold = assistantConfidenceThresholdIt->value.GetDouble();
-    }
-    if (maxNgramSizeItHasValue) {
-        request.maxNgramSize = maxNgramSizeIt->value.GetUint();
-    }
-    request.maxModelLength = maxModelLength;
-
-    // TODO: logit_bias
-    // TODO: top_logprobs
-    // TODO: response_format
-    // TODO: tools
-    // TODO: tool_choice
-    // TODO: user
-    // TODO: function_call (deprecated)
-    // TODO: functions (deprecated)
-    return absl::OkStatus();
-}
-
-std::optional<std::string> OpenAIChatCompletionsHandler::getPrompt() const { return request.prompt; }
-std::optional<int> OpenAIChatCompletionsHandler::getNumReturnSequences() const { return request.numReturnSequences; }
-StreamOptions OpenAIChatCompletionsHandler::getStreamOptions() const { return request.streamOptions; }
-
-bool OpenAIChatCompletionsHandler::isStream() const { return request.stream; }
-Endpoint OpenAIChatCompletionsHandler::getEndpoint() const { return endpoint; }
-std::string OpenAIChatCompletionsHandler::getModel() const { return request.model; }
-std::string OpenAIChatCompletionsHandler::getToolChoice() const { return request.toolChoice; }
-const std::unique_ptr<OutputParser>& OpenAIChatCompletionsHandler::getOutputParser() const { return outputParser; }
-
-void OpenAIChatCompletionsHandler::setPromptTokensUsage(size_t promptTokens) {
-    usage.promptTokens = promptTokens;
-}
-
-void OpenAIChatCompletionsHandler::setCompletionTokensUsage(size_t completionTokens) {
-    usage.completionTokens = completionTokens;
-}
-
-void OpenAIChatCompletionsHandler::incrementProcessedTokens(size_t numTokens) {
-    processedTokens += numTokens;
-    if (!request.echo || processedTokens > usage.promptTokens)
-        usage.completionTokens += numTokens;
-}
-
-absl::Status OpenAIChatCompletionsHandler::parseRequest(std::optional<uint32_t> maxTokensLimit, uint32_t bestOfLimit, std::optional<uint32_t> maxModelLength, std::optional<std::string> allowedLocalMediaPath, std::optional<std::vector<std::string>> allowedMediaDomains) {
-    absl::Status status = parseCommonPart(maxTokensLimit, bestOfLimit, maxModelLength);
-    if (status != absl::OkStatus())
-        return status;
-    if (endpoint == Endpoint::COMPLETIONS)
-        status = parseCompletionsPart();
-    else if (endpoint == Endpoint::RESPONSES)
-        status = parseResponsesPart(maxTokensLimit, allowedLocalMediaPath, allowedMediaDomains);
-    else
-        status = parseChatCompletionsPart(maxTokensLimit, allowedLocalMediaPath, allowedMediaDomains);
-
-    return status;
-}
-
-void updateUsage(CompletionUsageStatistics& usage, const std::vector<int64_t>& generatedIds, bool echoPrompt) {
-    OVMS_PROFILE_FUNCTION();
-    usage.completionTokens += generatedIds.size();
-    if (echoPrompt)
-        usage.completionTokens -= usage.promptTokens;
-}
-
-static std::optional<std::string> mapFinishReason(ov::genai::GenerationFinishReason finishReason, bool hasToolCalls) {
-    // GenerationFinishReason::TOOL_CALLS is not available in GenAI yet.
-    // Use feature detection based on presence of tool calls as a workaround until GenAI exposes TOOL_CALLS.
-    if (hasToolCalls && finishReason == ov::genai::GenerationFinishReason::STOP) {
-        return "tool_calls";
-    }
-    switch (finishReason) {
-    case ov::genai::GenerationFinishReason::STOP:
-        return "stop";
-    case ov::genai::GenerationFinishReason::LENGTH:
-        return "length";
-    default:
-        return std::nullopt;
-    }
-}
-
-static bool hasToolCallsInStreamingDelta(const rapidjson::Document& delta) {
-    if (!delta.HasMember("delta") || !delta["delta"].IsObject()) {
-        return false;
-    }
-    const auto& deltaObj = delta["delta"];
-    return deltaObj.HasMember("tool_calls") && deltaObj["tool_calls"].IsArray();
-}
-
-ParsedOutput OpenAIChatCompletionsHandler::parseOutputIfNeeded(const std::vector<int64_t>& generatedIds) {
-    OVMS_PROFILE_FUNCTION();
-    ParsedOutput parsedOutput;
-    if ((endpoint != Endpoint::CHAT_COMPLETIONS && endpoint != Endpoint::RESPONSES) || outputParser == nullptr) {
-        parsedOutput.content = this->tokenizer.decode(generatedIds);
-    } else {
-        parsedOutput = outputParser->parse(generatedIds, this->areToolsAvailable());
-    }
-    return parsedOutput;
-}
+// --- Unary response serialization ---
 
 std::string OpenAIChatCompletionsHandler::serializeUnaryResponse(const std::vector<ov::genai::GenerationOutput>& generationOutputs) {
     OVMS_PROFILE_FUNCTION();
-    if (endpoint == Endpoint::RESPONSES) {
-        std::vector<ParsedOutput> parsedOutputs;
-        usage.completionTokens = 0;
-        constexpr bool echo = false;  // echo is not supported in Responses API
-        ov::genai::GenerationFinishReason responsesFinishReason = ov::genai::GenerationFinishReason::STOP;
-        for (const ov::genai::GenerationOutput& generationOutput : generationOutputs) {
-            updateUsage(usage, generationOutput.generated_ids, echo);
-            parsedOutputs.push_back(parseOutputIfNeeded(generationOutput.generated_ids));
-            if (generationOutput.finish_reason == ov::genai::GenerationFinishReason::LENGTH) {
-                responsesFinishReason = ov::genai::GenerationFinishReason::LENGTH;
-            }
-        }
-        return serializeResponsesUnaryResponse(parsedOutputs, responsesFinishReason);
-    }
 
     OpenAiJsonResponse jsonResponse;
     jsonResponse.StartObject();
@@ -1900,13 +472,6 @@ std::string OpenAIChatCompletionsHandler::serializeUnaryResponse(ov::genai::Enco
     OVMS_PROFILE_FUNCTION();
     usage.promptTokens = results.perf_metrics.get_num_input_tokens();
     usage.completionTokens = results.perf_metrics.get_num_generated_tokens();
-    if (endpoint == Endpoint::RESPONSES) {
-        std::vector<ParsedOutput> parsedOutputs;
-        for (const auto& tokens : results.tokens) {
-            parsedOutputs.push_back(parseOutputIfNeeded(tokens));
-        }
-        return serializeResponsesUnaryResponse(parsedOutputs);
-    }
 
     OpenAiJsonResponse jsonResponse;
     jsonResponse.StartObject();
@@ -1966,33 +531,6 @@ std::string OpenAIChatCompletionsHandler::serializeUnaryResponse(ov::genai::VLMD
     OVMS_PROFILE_FUNCTION();
     usage.promptTokens = results.perf_metrics.get_num_input_tokens();
     usage.completionTokens = results.perf_metrics.get_num_generated_tokens();
-    if (endpoint == Endpoint::RESPONSES) {
-        // Usage is already correctly set from perf_metrics above — no need for updateUsage.
-        std::vector<ParsedOutput> parsedOutputs;
-        for (const std::string& text : results.texts) {
-            if (outputParser != nullptr) {
-                // Same workaround as in chat completions
-                auto result = tokenizer.encode(text);
-                auto& input_ids = result.input_ids;
-                if (input_ids.get_shape().size() != 2)
-                    throw std::runtime_error("input_ids should have 2 dimensions");
-                if (input_ids.get_shape()[0] != 1)
-                    throw std::runtime_error("input_ids should have 1 batch size");
-                if (input_ids.get_element_type() != ov::element::i64)
-                    throw std::runtime_error("input_ids should have i64 element type");
-
-                int64_t* inputIdsData = reinterpret_cast<int64_t*>(input_ids.data());
-                std::vector<int64_t> generatedTokens(inputIdsData, inputIdsData + input_ids.get_shape()[1]);
-                parsedOutputs.push_back(parseOutputIfNeeded(generatedTokens));
-            } else {
-                // Fast path: no output parser, use decoded text directly.
-                ParsedOutput output;
-                output.content = text;
-                parsedOutputs.push_back(std::move(output));
-            }
-        }
-        return serializeResponsesUnaryResponse(parsedOutputs);
-    }
 
     OpenAiJsonResponse jsonResponse;
     jsonResponse.StartObject();
@@ -2066,511 +604,10 @@ std::string OpenAIChatCompletionsHandler::serializeUnaryResponse(ov::genai::VLMD
     return jsonResponse.ToString();
 }
 
-void OpenAIChatCompletionsHandler::writeEventHeader(Writer<StringBuffer>& writer, const char* eventType) {
-    writer.StartObject();
-    writer.String("type");
-    writer.String(eventType);
-    writer.String("sequence_number");
-    writer.Uint64(responsesState.sequenceNumber++);
-}
-
-void OpenAIChatCompletionsHandler::writeContentLocation(Writer<StringBuffer>& writer, const std::string& itemId, uint64_t outputIndex) {
-    writer.String("output_index");
-    writer.Uint64(outputIndex);
-    writer.String("content_index");
-    writer.Uint64(0);
-    writer.String("item_id");
-    writer.String(itemId.c_str());
-}
-
-void OpenAIChatCompletionsHandler::writeReasoningLocation(Writer<StringBuffer>& writer, const std::string& itemId) {
-    writer.String("output_index");
-    writer.Uint64(0);
-    writer.String("summary_index");
-    writer.Uint64(0);
-    writer.String("item_id");
-    writer.String(itemId.c_str());
-}
-
-std::string OpenAIChatCompletionsHandler::serializeResponseCreatedEvent(const std::string& responseId, int64_t createdAt) {
-    StringBuffer buffer;
-    Writer<StringBuffer> writer(buffer);
-    writeEventHeader(writer, "response.created");
-    writer.String("response");
-    serializeResponsesResponseObject(writer, responseId, createdAt, "in_progress", "", false);
-    writer.EndObject();
-    return buffer.GetString();
-}
-
-std::string OpenAIChatCompletionsHandler::serializeResponseInProgressEvent(const std::string& responseId, int64_t createdAt) {
-    StringBuffer buffer;
-    Writer<StringBuffer> writer(buffer);
-    writeEventHeader(writer, "response.in_progress");
-    writer.String("response");
-    serializeResponsesResponseObject(writer, responseId, createdAt, "in_progress", "", false);
-    writer.EndObject();
-    return buffer.GetString();
-}
-
-std::string OpenAIChatCompletionsHandler::serializeOutputItemAddedEvent(const std::string& outputItemId, uint64_t outputIndex) {
-    StringBuffer buffer;
-    Writer<StringBuffer> writer(buffer);
-    writeEventHeader(writer, "response.output_item.added");
-    writer.String("output_index");
-    writer.Uint64(outputIndex);
-    writer.String("item");
-    serializeResponsesOutputItem(writer, outputItemId, "", "in_progress");
-    writer.EndObject();
-    return buffer.GetString();
-}
-
-std::string OpenAIChatCompletionsHandler::serializeContentPartAddedEvent(const std::string& outputItemId, uint64_t outputIndex) {
-    StringBuffer buffer;
-    Writer<StringBuffer> writer(buffer);
-    writeEventHeader(writer, "response.content_part.added");
-    writeContentLocation(writer, outputItemId, outputIndex);
-    writer.String("part");
-    serializeOutputTextPart(writer, "");
-    writer.EndObject();
-    return buffer.GetString();
-}
-
-std::string OpenAIChatCompletionsHandler::serializeOutputTextDeltaEvent(const std::string& outputItemId, const std::string& delta, uint64_t outputIndex) {
-    StringBuffer buffer;
-    Writer<StringBuffer> writer(buffer);
-    writeEventHeader(writer, "response.output_text.delta");
-    writeContentLocation(writer, outputItemId, outputIndex);
-    writer.String("delta");
-    writer.String(delta.c_str());
-    // TODO: logprobs not supported
-    writer.EndObject();
-    return buffer.GetString();
-}
-
-std::string OpenAIChatCompletionsHandler::serializeOutputTextDoneEvent(const std::string& outputItemId, uint64_t outputIndex) {
-    StringBuffer buffer;
-    Writer<StringBuffer> writer(buffer);
-    writeEventHeader(writer, "response.output_text.done");
-    writeContentLocation(writer, outputItemId, outputIndex);
-    writer.String("text");
-    writer.String(responsesState.outputText.c_str());
-    // TODO: logprobs not supported
-    writer.EndObject();
-    return buffer.GetString();
-}
-
-std::string OpenAIChatCompletionsHandler::serializeContentPartDoneEvent(const std::string& outputItemId, uint64_t outputIndex) {
-    StringBuffer buffer;
-    Writer<StringBuffer> writer(buffer);
-    writeEventHeader(writer, "response.content_part.done");
-    writeContentLocation(writer, outputItemId, outputIndex);
-    writer.String("part");
-    serializeOutputTextPart(writer, responsesState.outputText);
-    writer.EndObject();
-    return buffer.GetString();
-}
-
-std::string OpenAIChatCompletionsHandler::serializeOutputItemDoneEvent(const std::string& outputItemId, ov::genai::GenerationFinishReason finishReason, uint64_t outputIndex) {
-    const std::string itemStatus = (finishReason == ov::genai::GenerationFinishReason::LENGTH) ? "incomplete" : "completed";
-    StringBuffer buffer;
-    Writer<StringBuffer> writer(buffer);
-    writeEventHeader(writer, "response.output_item.done");
-    writer.String("output_index");
-    writer.Uint64(outputIndex);
-    writer.String("item");
-    serializeResponsesOutputItem(writer, outputItemId, responsesState.outputText, itemStatus);
-    writer.EndObject();
-    return buffer.GetString();
-}
-
-std::string OpenAIChatCompletionsHandler::serializeResponseCompletedEvent(const std::string& responseId, int64_t createdAt, ov::genai::GenerationFinishReason finishReason) {
-    const bool isIncomplete = (finishReason == ov::genai::GenerationFinishReason::LENGTH);
-    const std::string responseStatus = isIncomplete ? "incomplete" : "completed";
-    const char* eventType = isIncomplete ? "response.incomplete" : "response.completed";
-    std::optional<std::string> incompleteReason = isIncomplete ? std::optional<std::string>("max_tokens") : std::nullopt;
-    StringBuffer buffer;
-    Writer<StringBuffer> writer(buffer);
-    writeEventHeader(writer, eventType);
-    writer.String("response");
-    serializeResponsesResponseObject(writer, responseId, createdAt, responseStatus, responsesState.outputText, true, incompleteReason);
-    writer.EndObject();
-    return buffer.GetString();
-}
-
-std::string OpenAIChatCompletionsHandler::serializeResponseFailedEventBody(const std::string& responseId, int64_t createdAt, const std::string& errorMessage, ResponsesErrorCode errorCode) {
-    StringBuffer buffer;
-    Writer<StringBuffer> writer(buffer);
-    writeEventHeader(writer, "response.failed");
-    writer.String("response");
-    serializeResponsesResponseObject(writer, responseId, createdAt, "failed", responsesState.outputText, false,
-        std::nullopt, errorMessage, errorCode);
-    writer.EndObject();
-    return buffer.GetString();
-}
-
-std::string OpenAIChatCompletionsHandler::serializeReasoningOutputItemAddedEvent(const std::string& reasoningItemId) {
-    StringBuffer buffer;
-    Writer<StringBuffer> writer(buffer);
-    writeEventHeader(writer, "response.output_item.added");
-    writer.String("output_index");
-    writer.Uint64(0);
-    writer.String("item");
-    writer.StartObject();
-    writer.String("id");
-    writer.String(reasoningItemId.c_str());
-    writer.String("type");
-    writer.String("reasoning");
-    writer.String("summary");
-    writer.StartArray();
-    writer.EndArray();
-    writer.EndObject();
-    writer.EndObject();
-    return buffer.GetString();
-}
-
-std::string OpenAIChatCompletionsHandler::serializeReasoningSummaryPartAddedEvent(const std::string& reasoningItemId) {
-    StringBuffer buffer;
-    Writer<StringBuffer> writer(buffer);
-    writeEventHeader(writer, "response.reasoning_summary_part.added");
-    writeReasoningLocation(writer, reasoningItemId);
-    writer.String("part");
-    writer.StartObject();
-    writer.String("type");
-    writer.String("summary_text");
-    writer.String("text");
-    writer.String("");
-    writer.EndObject();
-    writer.EndObject();
-    return buffer.GetString();
-}
-
-std::string OpenAIChatCompletionsHandler::serializeReasoningSummaryTextDeltaEvent(const std::string& reasoningItemId, const std::string& delta) {
-    StringBuffer buffer;
-    Writer<StringBuffer> writer(buffer);
-    writeEventHeader(writer, "response.reasoning_summary_text.delta");
-    writeReasoningLocation(writer, reasoningItemId);
-    writer.String("delta");
-    writer.String(delta.c_str());
-    writer.EndObject();
-    return buffer.GetString();
-}
-
-std::string OpenAIChatCompletionsHandler::serializeReasoningSummaryTextDoneEvent(const std::string& reasoningItemId) {
-    StringBuffer buffer;
-    Writer<StringBuffer> writer(buffer);
-    writeEventHeader(writer, "response.reasoning_summary_text.done");
-    writeReasoningLocation(writer, reasoningItemId);
-    writer.String("text");
-    writer.String(responsesState.reasoningText.c_str());
-    writer.EndObject();
-    return buffer.GetString();
-}
-
-std::string OpenAIChatCompletionsHandler::serializeReasoningSummaryPartDoneEvent(const std::string& reasoningItemId) {
-    StringBuffer buffer;
-    Writer<StringBuffer> writer(buffer);
-    writeEventHeader(writer, "response.reasoning_summary_part.done");
-    writeReasoningLocation(writer, reasoningItemId);
-    writer.String("part");
-    writer.StartObject();
-    writer.String("type");
-    writer.String("summary_text");
-    writer.String("text");
-    writer.String(responsesState.reasoningText.c_str());
-    writer.EndObject();
-    writer.EndObject();
-    return buffer.GetString();
-}
-
-std::string OpenAIChatCompletionsHandler::serializeReasoningOutputItemDoneEvent(const std::string& reasoningItemId) {
-    StringBuffer buffer;
-    Writer<StringBuffer> writer(buffer);
-    writeEventHeader(writer, "response.output_item.done");
-    writer.String("output_index");
-    writer.Uint64(0);
-    writer.String("item");
-    writer.StartObject();
-    writer.String("id");
-    writer.String(reasoningItemId.c_str());
-    writer.String("type");
-    writer.String("reasoning");
-    writer.String("summary");
-    writer.StartArray();
-    writer.StartObject();
-    writer.String("type");
-    writer.String("summary_text");
-    writer.String("text");
-    writer.String(responsesState.reasoningText.c_str());
-    writer.EndObject();
-    writer.EndArray();
-    writer.EndObject();
-    writer.EndObject();
-    return buffer.GetString();
-}
-
-std::string OpenAIChatCompletionsHandler::serializeFunctionCallOutputItemAddedEvent(const ToolCall& toolCall, uint64_t outputIndex) {
-    StringBuffer buffer;
-    Writer<StringBuffer> writer(buffer);
-    writeEventHeader(writer, "response.output_item.added");
-    writer.String("output_index");
-    writer.Uint64(outputIndex);
-    writer.String("item");
-    writer.StartObject();
-    writer.String("id");
-    writer.String(toolCall.id.c_str());
-    writer.String("type");
-    writer.String("function_call");
-    writer.String("status");
-    writer.String("in_progress");
-    writer.String("call_id");
-    writer.String(toolCall.id.c_str());
-    writer.String("name");
-    writer.String(toolCall.name.c_str());
-    writer.String("arguments");
-    writer.String("");
-    writer.EndObject();
-    writer.EndObject();
-    return buffer.GetString();
-}
-
-std::string OpenAIChatCompletionsHandler::serializeFunctionCallArgumentsDeltaEvent(const std::string& callId, const std::string& delta, uint64_t outputIndex) {
-    StringBuffer buffer;
-    Writer<StringBuffer> writer(buffer);
-    writeEventHeader(writer, "response.function_call_arguments.delta");
-    writer.String("item_id");
-    writer.String(callId.c_str());
-    writer.String("output_index");
-    writer.Uint64(outputIndex);
-    writer.String("call_id");
-    writer.String(callId.c_str());
-    writer.String("delta");
-    writer.String(delta.c_str());
-    writer.EndObject();
-    return buffer.GetString();
-}
-
-std::string OpenAIChatCompletionsHandler::serializeFunctionCallArgumentsDoneEvent(const ToolCall& toolCall, uint64_t outputIndex) {
-    StringBuffer buffer;
-    Writer<StringBuffer> writer(buffer);
-    writeEventHeader(writer, "response.function_call_arguments.done");
-    writer.String("item_id");
-    writer.String(toolCall.id.c_str());
-    writer.String("output_index");
-    writer.Uint64(outputIndex);
-    writer.String("call_id");
-    writer.String(toolCall.id.c_str());
-    writer.String("arguments");
-    writer.String(toolCall.arguments.c_str());
-    writer.EndObject();
-    return buffer.GetString();
-}
-
-std::string OpenAIChatCompletionsHandler::serializeFunctionCallOutputItemDoneEvent(const ToolCall& toolCall, ov::genai::GenerationFinishReason finishReason, uint64_t outputIndex) {
-    const std::string itemStatus = (finishReason == ov::genai::GenerationFinishReason::LENGTH) ? "incomplete" : "completed";
-    StringBuffer buffer;
-    Writer<StringBuffer> writer(buffer);
-    writeEventHeader(writer, "response.output_item.done");
-    writer.String("output_index");
-    writer.Uint64(outputIndex);
-    writer.String("item");
-    writer.StartObject();
-    writer.String("id");
-    writer.String(toolCall.id.c_str());
-    writer.String("type");
-    writer.String("function_call");
-    writer.String("status");
-    writer.String(itemStatus.c_str());
-    writer.String("call_id");
-    writer.String(toolCall.id.c_str());
-    writer.String("name");
-    writer.String(toolCall.name.c_str());
-    writer.String("arguments");
-    writer.String(toolCall.arguments.c_str());
-    writer.EndObject();
-    writer.EndObject();
-    return buffer.GetString();
-}
-
-std::string OpenAIChatCompletionsHandler::serializeResponsesStreamingInitEvents() {
-    const auto createdAt = std::chrono::duration_cast<std::chrono::microseconds>(created.time_since_epoch()).count();
-    const std::string responseId = "resp-" + std::to_string(createdAt);
-    const std::string outputItemId = "msg-0";
-
-    std::vector<std::string> events;
-
-    events.emplace_back(serializeResponseCreatedEvent(responseId, createdAt));
-    events.emplace_back(serializeResponseInProgressEvent(responseId, createdAt));
-
-    // When outputParser is present, defer output item events until first chunk
-    // because reasoning items need to come before message items
-    if (outputParser == nullptr) {
-        events.emplace_back(serializeOutputItemAddedEvent(outputItemId));
-        events.emplace_back(serializeContentPartAddedEvent(outputItemId));
-        responsesState.messageInitialized = true;
-    }
-
-    responsesState.initialized = true;
-
-    std::stringstream ss;
-    ss << events.front();
-    for (size_t i = 1; i < events.size(); ++i) {
-        ss << "\n\ndata: " << events[i];
-    }
-    return ss.str();
-}
+// --- Streaming serialization ---
 
 std::string OpenAIChatCompletionsHandler::serializeStreamingChunk(const std::string& chunkResponse, ov::genai::GenerationFinishReason finishReason) {
     OVMS_PROFILE_FUNCTION();
-    if (endpoint == Endpoint::RESPONSES) {
-        const auto createdAt = std::chrono::duration_cast<std::chrono::microseconds>(created.time_since_epoch()).count();
-        const std::string responseId = "resp-" + std::to_string(createdAt);
-        const std::string outputItemId = "msg-0";
-        const std::string reasoningItemId = "rs-0";
-
-        std::vector<std::string> events;
-        if (!responsesState.initialized) {
-            // Fallback: if init events were not sent earlier, emit them now
-            std::string initEvents = serializeResponsesStreamingInitEvents();
-            if (!initEvents.empty()) {
-                events.emplace_back(std::move(initEvents));
-            }
-        }
-
-        if (outputParser != nullptr) {
-            // Use output parser to separate reasoning from content
-            std::optional<Document> delta = outputParser->parseChunk(chunkResponse, areToolsAvailable(), finishReason);
-
-            if (delta.has_value() && delta->HasMember("delta") && (*delta)["delta"].IsObject()) {
-                const auto& deltaObj = (*delta)["delta"];
-                if (deltaObj.HasMember("reasoning_content") && deltaObj["reasoning_content"].IsString()) {
-                    // Reasoning chunk
-                    if (!responsesState.reasoningInitialized) {
-                        events.emplace_back(serializeReasoningOutputItemAddedEvent(reasoningItemId));
-                        events.emplace_back(serializeReasoningSummaryPartAddedEvent(reasoningItemId));
-                        responsesState.reasoningInitialized = true;
-                    }
-                    const std::string reasoningText = deltaObj["reasoning_content"].GetString();
-                    responsesState.reasoningText += reasoningText;
-                    events.emplace_back(serializeReasoningSummaryTextDeltaEvent(reasoningItemId, reasoningText));
-                } else if (deltaObj.HasMember("content") && deltaObj["content"].IsString()) {
-                    // Content chunk - close reasoning if it was active, init message if needed
-                    if (responsesState.reasoningInitialized && !responsesState.reasoningCompleted) {
-                        events.emplace_back(serializeReasoningSummaryTextDoneEvent(reasoningItemId));
-                        events.emplace_back(serializeReasoningSummaryPartDoneEvent(reasoningItemId));
-                        events.emplace_back(serializeReasoningOutputItemDoneEvent(reasoningItemId));
-                        responsesState.reasoningCompleted = true;
-                    }
-                    const uint64_t msgIdx = responsesState.reasoningInitialized ? 1 : 0;
-                    if (!responsesState.messageInitialized) {
-                        events.emplace_back(serializeOutputItemAddedEvent(outputItemId, msgIdx));
-                        events.emplace_back(serializeContentPartAddedEvent(outputItemId, msgIdx));
-                        responsesState.messageInitialized = true;
-                    }
-                    const std::string contentText = deltaObj["content"].GetString();
-                    responsesState.outputText += contentText;
-                    events.emplace_back(serializeOutputTextDeltaEvent(outputItemId, contentText, msgIdx));
-                } else if (deltaObj.HasMember("tool_calls") && deltaObj["tool_calls"].IsArray()) {
-                    // Tool call chunk - close reasoning if active
-                    if (responsesState.reasoningInitialized && !responsesState.reasoningCompleted) {
-                        events.emplace_back(serializeReasoningSummaryTextDoneEvent(reasoningItemId));
-                        events.emplace_back(serializeReasoningSummaryPartDoneEvent(reasoningItemId));
-                        events.emplace_back(serializeReasoningOutputItemDoneEvent(reasoningItemId));
-                        responsesState.reasoningCompleted = true;
-                    }
-                    const auto& toolCallsArr = deltaObj["tool_calls"];
-                    for (rapidjson::SizeType i = 0; i < toolCallsArr.Size(); ++i) {
-                        const auto& tc = toolCallsArr[i];
-                        int tcIndex = tc.HasMember("index") ? tc["index"].GetInt() : 0;
-                        // Determine the output index for this tool call
-                        const uint64_t baseIdx = responsesState.reasoningInitialized ? 1 : 0;
-                        const uint64_t tcOutputIdx = baseIdx + static_cast<uint64_t>(tcIndex);
-                        // Determine if this is a new tool call (has function name)
-                        bool isNewToolCall = false;
-                        std::string funcName;
-                        std::string tcId;
-                        std::string argDelta;
-                        if (tc.HasMember("function") && tc["function"].IsObject()) {
-                            const auto& funcObj = tc["function"];
-                            if (funcObj.HasMember("name") && funcObj["name"].IsString()) {
-                                funcName = funcObj["name"].GetString();
-                                isNewToolCall = true;
-                            }
-                            if (funcObj.HasMember("arguments") && funcObj["arguments"].IsString()) {
-                                argDelta = funcObj["arguments"].GetString();
-                            }
-                        }
-                        if (tc.HasMember("id") && tc["id"].IsString()) {
-                            tcId = tc["id"].GetString();
-                        }
-                        if (isNewToolCall) {
-                            // Ensure we have enough entries in our tracking vector
-                            while (static_cast<int>(responsesState.toolCalls.size()) <= tcIndex) {
-                                responsesState.toolCalls.push_back(ToolCall{});
-                            }
-                            responsesState.toolCalls[tcIndex].id = tcId;
-                            responsesState.toolCalls[tcIndex].name = funcName;
-                            responsesState.toolCalls[tcIndex].arguments = "";
-                            events.emplace_back(serializeFunctionCallOutputItemAddedEvent(responsesState.toolCalls[tcIndex], tcOutputIdx));
-                        }
-                        if (!argDelta.empty() && static_cast<int>(responsesState.toolCalls.size()) > tcIndex) {
-                            responsesState.toolCalls[tcIndex].arguments += argDelta;
-                            events.emplace_back(serializeFunctionCallArgumentsDeltaEvent(responsesState.toolCalls[tcIndex].id, argDelta, tcOutputIdx));
-                        }
-                    }
-                }
-            }
-            // If delta is nullopt, the parser is accumulating tag tokens - skip
-        } else {
-            // No parser - pass through raw text
-            if (!chunkResponse.empty()) {
-                responsesState.outputText += chunkResponse;
-                events.emplace_back(serializeOutputTextDeltaEvent(outputItemId, chunkResponse));
-            }
-        }
-
-        if (finishReason != ov::genai::GenerationFinishReason::NONE) {
-            // Close any open reasoning that wasn't closed by content transition
-            if (responsesState.reasoningInitialized && !responsesState.reasoningCompleted) {
-                events.emplace_back(serializeReasoningSummaryTextDoneEvent(reasoningItemId));
-                events.emplace_back(serializeReasoningSummaryPartDoneEvent(reasoningItemId));
-                events.emplace_back(serializeReasoningOutputItemDoneEvent(reasoningItemId));
-                responsesState.reasoningCompleted = true;
-            }
-            // Emit done events for any streaming tool calls
-            if (!responsesState.toolCalls.empty()) {
-                const uint64_t baseIdx = responsesState.reasoningInitialized ? 1 : 0;
-                for (size_t i = 0; i < responsesState.toolCalls.size(); ++i) {
-                    const uint64_t tcOutputIdx = baseIdx + static_cast<uint64_t>(i);
-                    events.emplace_back(serializeFunctionCallArgumentsDoneEvent(responsesState.toolCalls[i], tcOutputIdx));
-                    events.emplace_back(serializeFunctionCallOutputItemDoneEvent(responsesState.toolCalls[i], finishReason, tcOutputIdx));
-                }
-            }
-            // Only emit message item if content was produced or no tool calls were generated
-            if (!responsesState.outputText.empty() || responsesState.toolCalls.empty()) {
-                const uint64_t msgIdx = (responsesState.reasoningInitialized ? 1 : 0) + responsesState.toolCalls.size();
-                if (!responsesState.messageInitialized) {
-                    events.emplace_back(serializeOutputItemAddedEvent(outputItemId, msgIdx));
-                    events.emplace_back(serializeContentPartAddedEvent(outputItemId, msgIdx));
-                    responsesState.messageInitialized = true;
-                }
-                events.emplace_back(serializeOutputTextDoneEvent(outputItemId, msgIdx));
-                events.emplace_back(serializeContentPartDoneEvent(outputItemId, msgIdx));
-                events.emplace_back(serializeOutputItemDoneEvent(outputItemId, finishReason, msgIdx));
-            }
-            events.emplace_back(serializeResponseCompletedEvent(responseId, createdAt, finishReason));
-        }
-
-        if (events.empty()) {
-            return "";
-        }
-
-        std::stringstream ss;
-        ss << events.front();
-        for (size_t i = 1; i < events.size(); ++i) {
-            ss << "\n\ndata: " << events[i];
-        }
-        return ss.str();
-    }
 
     Document doc;
     doc.SetObject();
@@ -2662,33 +699,8 @@ std::string OpenAIChatCompletionsHandler::serializeStreamingChunk(const std::str
     return buffer.GetString();
 }
 
-std::string OpenAIChatCompletionsHandler::serializeResponsesFailedEvent(const std::string& errorMessage, ResponsesErrorCode errorCode) {
-    const auto createdAt = std::chrono::duration_cast<std::chrono::microseconds>(created.time_since_epoch()).count();
-    const std::string responseId = "resp-" + std::to_string(createdAt);
-
-    std::vector<std::string> events;
-    if (!responsesState.initialized) {
-        std::string initEvents = serializeResponsesStreamingInitEvents();
-        if (!initEvents.empty()) {
-            events.emplace_back(std::move(initEvents));
-        }
-    }
-
-    events.emplace_back(serializeResponseFailedEventBody(responseId, createdAt, errorMessage, errorCode));
-
-    std::stringstream ss;
-    ss << events.front();
-    for (size_t i = 1; i < events.size(); ++i) {
-        ss << "\n\ndata: " << events[i];
-    }
-    return ss.str();
-}
-
 std::string OpenAIChatCompletionsHandler::serializeStreamingUsageChunk() {
     OVMS_PROFILE_FUNCTION();
-    if (endpoint == Endpoint::RESPONSES) {
-        return "";
-    }
     StringBuffer buffer;
     Writer<StringBuffer> writer(buffer);
