+++ src/llm/servable.cpp
@@ -34,6 +34,7 @@
 #include "../mediapipe_internal/mediapipe_utils.hpp"
 #include "../profiler.hpp"
 #include "apis/openai_completions.hpp"
+#include "apis/openai_responses.hpp"
 #include "servable.hpp"
 #include "text_utils.hpp"
 #include "../tokenize/tokenize_parser.hpp"
@@ -110,12 +111,21 @@ absl::Status GenAiServable::processTokenizeRequest(std::shared_ptr<GenAiServable
 
 absl::Status GenAiServable::parseRequest(std::shared_ptr<GenAiServableExecutionContext>& executionContext) {
     try {
-        executionContext->apiHandler = std::make_shared<OpenAIChatCompletionsHandler>(*executionContext->payload.parsedJson,
-            executionContext->endpoint,
-            std::chrono::system_clock::now(),
-            getProperties()->tokenizer,
-            getProperties()->toolParserName,
-            getProperties()->reasoningParserName);
+        if (executionContext->endpoint == Endpoint::RESPONSES) {
+            executionContext->apiHandler = std::make_shared<OpenAIResponsesHandler>(*executionContext->payload.parsedJson,
+                executionContext->endpoint,
+                std::chrono::system_clock::now(),
+                getProperties()->tokenizer,
+                getProperties()->toolParserName,
+                getProperties()->reasoningParserName);
+        } else {
+            executionContext->apiHandler = std::make_shared<OpenAIChatCompletionsHandler>(*executionContext->payload.parsedJson,
+                executionContext->endpoint,
+                std::chrono::system_clock::now(),
+                getProperties()->tokenizer,
+                getProperties()->toolParserName,
+                getProperties()->reasoningParserName);
+        }
     } catch (const std::exception& e) {
         SPDLOG_LOGGER_ERROR(llm_calculator_logger, "Failed to create API handler: {}", e.what());
         return absl::InvalidArgumentError(std::string("Failed to create API handler: ") + e.what());
@@ -306,8 +316,8 @@ absl::Status GenAiServable::preparePartialResponse(std::shared_ptr<GenAiServable
     ov::genai::GenerationFinishReason finishReason = generationOutput.finish_reason;
     if (finishReason == ov::genai::GenerationFinishReason::NONE) {  // continue
         // For RESPONSES endpoint, always call serializeStreamingChunk so that
-        // initialization events (response.created, response.in_progress, etc.)
-        // are emitted immediately, even before the tokenizer produces text.
+        // output item initialization events (output_item.added, content_part.added)
+        // are emitted on the first chunk, even before the tokenizer produces text.
         if (lastTextChunk.size() > 0 || executionContext->apiHandler->getEndpoint() == Endpoint::RESPONSES) {
             std::string serializedChunk = executionContext->apiHandler->serializeStreamingChunk(lastTextChunk, finishReason);
             if (!serializedChunk.empty()) {
